### 2024-03-21 f3bab1b
typo in doc about single vs double precision
