+++ NLSE/nlse.py
@@ -2623,159 +2623,3 @@ class NLSE_1d_adim(NLSE_1d):
         )
         bq[..., 0] = 0
         return bq
-
-
-if __name__ == "__main__":
-
-    def normalize(arr: np.ndarray) -> np.ndarray:
-        """Normalize an array.
-
-        Args:
-            arr (np.ndarray): Array to normalize.
-
-        Returns:
-            np.ndarray: Normalized array.
-        """
-        return (arr - np.amin(arr)) / (np.amax(arr) - np.amin(arr))
-
-    def flatTop_tur(
-        sx: int,
-        sy: int,
-        length: int = 150,
-        width: int = 60,
-        k_counter: int = 81,
-        N_steps: int = 81,
-    ) -> np.ndarray:
-        """Collide two counterstreaming components.
-
-        Generates the phase mask to create two counterstreaming colliding
-        components.
-
-        Args:
-            sx (int): x Dimension of the mask : slm dimensions.
-            sy (int): y Dimension of the mask : slm dimensions.
-            length (int, optional): Length of the pattern. Defaults to 150.
-            width (int, optional): Width of the pattern. Defaults to 60.
-            k_counter (int, optional): Frequency of the blazed grating.
-            Defaults to 81.
-            N_steps (int, optional): Frequency of the vertical grating.
-            Defaults to 81.
-
-        Returns:
-            _type_: _description_
-        """
-        output = np.zeros((sy, sx))
-        Y, X = np.indices(output.shape)
-        output[abs(X - output.shape[1] // 2) < length / 2] = 1
-        output[abs(Y - output.shape[0] // 2) > width / 2] = 0
-
-        grating_axe = X
-        grating_axe = grating_axe % (sx / k_counter)
-        grating_axe += abs(np.amin(grating_axe))
-        grating_axe /= np.amax(grating_axe)
-
-        grating_axe[X > output.shape[1] // 2] *= -1
-        grating_axe[X > output.shape[1] // 2] += 1
-
-        grating_axe_vert = Y
-        grating_axe_vert = grating_axe_vert % (sy / N_steps)
-        grating_axe_vert = normalize(grating_axe_vert)
-
-        grating_axe = ((grating_axe + grating_axe_vert) % 1) * output
-        return grating_axe
-
-    def flatTop_super(
-        sx: int,
-        sy: int,
-        length: int = 150,
-        width: int = 60,
-        k_counter: int = 81,
-        N_steps: int = 81,
-    ) -> np.ndarray:
-        """Shear two components.
-
-        Generates the phase mask to create two counterstreaming shearing
-        components.
-
-        Args:
-            sx (int): x Dimension of the mask : slm dimensions.
-            sy (int): y Dimension of the mask : slm dimensions.
-            length (int, optional): Length of the pattern. Defaults to 150.
-            width (int, optional): Width of the pattern. Defaults to 60.
-            k_counter (int, optional): Frequency of the blazed grating.
-            Defaults to 81.
-            N_steps (int, optional): Frequency of the vertical grating.
[…]
