### 2024-03-21 #6509 @AlexBuccheri
Hi @RMeli , apologies but I don't have time for this one (and also don't know anything about the nonlinear Schrödinger equation 😅)
