### 2024-03-21 #6509 @RMeli
@taladjidi thank you for submitting to JOSS. While I look for reviewers, I started having a look at the code base and I noticed the following:

* In the test folder, I could not find any test framework; looking more closely, the tests don't seem to test anything (there are no assertions) and I was not able to locate any unit tests. Am I missing something?
* The `nlse.py` file is a single massive Python file of 2500+ lines of code. Calculations and analysis seems to be very tightly coupled within classes, and it seems that there is a lot of code duplication. There is also a lot of branching in the code (a lot of nested if/else), that makes the logic difficult to follow.

Could you please address the previous two points (proper testing, decoupling between calculations and analysis, removal of pervasive code duplication)?
