### 2024-03-22 #6509 @RMeli
> Trying to slowly work my way from "physicist" code to actual code

@taladjidi no worries, I totally get where you are coming from. If you need any suggestions do let me know.
