### 2024-03-26 10ba45c
Gitignore and remove static PRECISION
