+++ NLSE/nlse.py
@@ -10,21 +10,12 @@ import tqdm
 import matplotlib.pyplot as plt
 import numpy as np
 import pyfftw
-from scipy.constants import c, epsilon_0, hbar, atomic_mass
-from scipy.ndimage import zoom
+from scipy.constants import c, epsilon_0, hbar
 from scipy import signal
 from scipy import special
-from typing import Any
 from . import kernels_cpu
 
 BACKEND = "GPU"
-PRECISION = "single"
-if PRECISION == "double":
-    PRECISION_REAL = np.float64
-    PRECISION_COMPLEX = np.complex128
-else:
-    PRECISION_REAL = np.float32
-    PRECISION_COMPLEX = np.complex64
 
 if BACKEND == "GPU":
     try:
@@ -109,7 +100,7 @@ class NLSE:
             num=NX,
             endpoint=False,
             retstep=True,
-            dtype=PRECISION_REAL,
+            dtype=np.float32,
         )
         self.Y, self.delta_Y = np.linspace(
             -self.window / 2,
@@ -117,7 +108,7 @@ class NLSE:
             num=NY,
             endpoint=False,
             retstep=True,
-            dtype=PRECISION_REAL,
+            dtype=np.float32,
         )
 
         self.XX, self.YY = np.meshgrid(self.X, self.Y)
@@ -140,119 +131,7 @@ class NLSE:
             ] = np.nanmax(self.nl_profile[np.logical_not(np.isinf(self.nl_profile))])
             self.nl_profile /= self.nl_profile.sum()
         else:
-            self.nl_profile = np.ones((self.NY, self.NX), dtype=PRECISION_REAL)
-
-    def plot_2d(
-        self,
-        ax,
-        Z,
-        X,
-        AMP,
-        title,
-        cmap="viridis",
-        xlabel=r"$X$ (mm)",
-        ylabel=r"$Y (mm)$",
-        **kwargs,
-    ):
-        """Plots a 2d amplitude on an equidistant Z * X grid.
-
-        Args:
-            ax (matplotlib.Axes): The ax instance on which the plot will be drawn
-            Z (np.ndarray): the X axis values
-            X (np.ndarray): the Y axis values
-            AMP (np.ndarray): The 2D field to plot
-            title (str): Title of the plot
-            cmap (str, optional): Colormap. Defaults to 'viridis'.
-            label (str, optional): Label for the x axis. Defaults to r'$ (mm)'.
-            vmax (int, optional): Maximum value for cmap normalization. Defaults to 1.
-        """
-        im = ax.imshow(
-            AMP,
-            aspect="equal",
-            origin="lower",
-            extent=(Z[0], Z[-1], X[0], X[-1]),
-            cmap=cmap,
-            **kwargs,
-        )
-        ax.set_xlabel(xlabel)
-        ax.set_ylabel(ylabel)
-        ax.set_title(title)
-        plt.colorbar(im)
-        return
-
-    def plot_1d(
-        self, ax, T, labelT, AMP, labelAMP, PHASE, labelPHASE, Tmin, Tmax
-    ) -> None:
-        """Plots a 1D slice of a 2D field with amplitude and phase
-
-        Args:
-            ax (matplotlib.Axes): The ax instance on which the plot will be drawn
-            T (np.ndarray): The x axis
-            labelT (str):  x axis label
-            AMP (np.ndarray): The field slice
-            labelAMP (str): y axis label
-            PHASE (np.ndarray): The corresponding phase for the slice
-            labelPHASE (str): y axis label for the phase
-            Tmin (float): x axis left limit
-            Tmax (float): x axis right limit
[…]
