+++ NLSE/__init__.py
@@ -1,0 +1,5 @@
+from .nlse import NLSE
+from .nlse import NLSE_1d
+from .nlse import CNLSE
+from .nlse import CNLSE_1d
+from .nlse import GPE
