+++ tests/twod_uncoupled.py
@@ -1,4 +1,4 @@
-from NLSE.nlse import NLSE
+from NLSE import NLSE
 import numpy as np
 
 PRECISION_COMPLEX = np.complex64
@@ -10,7 +10,7 @@ def main():
     n2 = -1.6e-9
     waist = 2.23e-3
     waist2 = 70e-6
-    window = N * 5.5e-6
+    window = 4 * waist
     puiss = 1.05
     Isat = 10e4  # saturation intensity in W/m^2
     L = 10e-3
