### 2024-03-26 d188b01
Towards smarter inheritance
