+++ NLSE/nlse.py
@@ -10,7 +10,7 @@ import tqdm
 import matplotlib.pyplot as plt
 import numpy as np
 import pyfftw
-from scipy.constants import c, epsilon_0, hbar
+from scipy.constants import c, epsilon_0, hbar, atomic_mass
 from scipy import signal
 from scipy import special
 from . import kernels_cpu
@@ -72,12 +72,12 @@ class NLSE:
         # listof physical parameters
         self.backend = backend
         if self.backend == "GPU" and CUPY_AVAILABLE:
-            self.kernels = kernels_gpu
-            self.convolution = signal_cp.fftconvolve
+            self._kernels = kernels_gpu
+            self._convolution = signal_cp.fftconvolve
         else:
             self.backend = "CPU"
-            self.kernels = kernels_cpu
-            self.convolution = signal.fftconvolve
+            self._kernels = kernels_cpu
+            self._convolution = signal.fftconvolve
         self.n2 = n2
         self.V = V
         self.wl = wvl
@@ -110,6 +110,8 @@ class NLSE:
             retstep=True,
             dtype=np.float32,
         )
+        # define last axes for broadcasting operations
+        self._last_axes = (-2, -1)
 
         self.XX, self.YY = np.meshgrid(self.X, self.Y)
         # definition of the Fourier frequencies for the linear step
@@ -158,12 +160,10 @@ class NLSE:
             list: A list containing the FFT plans
         """
         if self.backend == "GPU" and CUPY_AVAILABLE:
-            # plan_fft = fftpack.get_fft_plan(
-            #     A, shape=A.shape, axes=(-2, -1), value_type='C2C')
             plan_fft = fftpack.get_fft_plan(
                 A,
                 shape=(A.shape[-2], A.shape[-1]),
-                axes=(-2, -1),
+                axes=self._last_axes,
                 value_type="C2C",
             )
             return [plan_fft]
@@ -180,14 +180,14 @@ class NLSE:
                 A,
                 direction="FFTW_FORWARD",
                 threads=multiprocessing.cpu_count(),
-                axes=(-2, -1),
+                axes=self._last_axes,
             )
             plan_ifft = pyfftw.FFTW(
                 A,
                 A,
                 direction="FFTW_BACKWARD",
                 threads=multiprocessing.cpu_count(),
-                axes=(-2, -1),
+                axes=self._last_axes,
             )
             with open("fft.wisdom", "wb") as file:
                 wisdom = pyfftw.export_wisdom()
@@ -223,11 +223,11 @@ class NLSE:
         if precision == "double":
             A_sq = A.real * A.real + A.imag * A.imag
             if self.nl_length > 0:
-                A_sq = self.convolution(
-                    A_sq, self.nl_profile, mode="same", axes=(-2, -1)
+                A_sq = self._convolution(
+                    A_sq, self.nl_profile, mode="same", axes=self._last_axes
                 )
             if V is None:
-                self.kernels.nl_prop_without_V(
+                self._kernels.nl_prop_without_V(
                     A,
                     A_sq,
                     self.delta_z / 2,
@@ -236,7 +236,7 @@ class NLSE:
                     2 * self.I_sat / (epsilon_0 * c),
                 )
             else:
-                self.self.kernels.nl_prop(
+                self.self._kernels.nl_prop(
                     A,
                     A_sq,
                     self.delta_z / 2,
@@ -258,10 +258,12 @@ class NLSE:
             plan_ifft(input_array=A, output_array=A, normalise_idft=True)
         A_sq = A.real * A.real + A.imag * A.imag
         if self.nl_length > 0:
[…]
