### 2024-03-26 366f642
further inheritance
