### 2024-03-26 cc6665d
inheritance simplification
