+++ NLSE/nlse.py
@@ -159,7 +159,6 @@ class NLSE:
         if self.backend == "GPU" and CUPY_AVAILABLE:
             plan_fft = fftpack.get_fft_plan(
                 A,
-                shape=(A.shape[-2], A.shape[-1]),
                 axes=self._last_axes,
                 value_type="C2C",
             )
@@ -248,7 +247,7 @@ class NLSE:
             cp.multiply(A, propagator, out=A)
             plan_fft.fft(A, A, cp.cuda.cufft.CUFFT_INVERSE)
             # fft normalization
-            A /= A.shape[-2] * A.shape[-1]
+            A /= np.prod(A.shape[self._last_axes[0] :])
         else:
             plan_fft(input_array=A, output_array=A)
             np.multiply(A, propagator, out=A)
@@ -427,7 +426,9 @@ class NLSE:
         Returns:
             np.ndarray: Propagated field in proper units V/m
         """
-        assert E_in.shape[-2] == self.NY and E_in.shape[-1] == self.NX
+        assert (
+            E_in.shape[self._last_axes[0] :] == self.XX.shape[self._last_axes[0] :]
+        ), "Shape mismatch"
         assert E_in.dtype in [
             np.complex64,
             np.complex128,
@@ -535,6 +536,7 @@ class NLSE_1d(NLSE):
             wvl=wvl,
             backend=backend,
         )
+        self._last_axes = (-1,)
         self.nl_profile = self.nl_profile[0]
         self.nl_profile /= self.nl_profile.sum()
 
@@ -554,277 +556,30 @@ class NLSE_1d(NLSE):
         else:
             return propagator
 
-    def build_fft_plan(self, A: np.ndarray) -> list:
-        """Builds the FFT plan objects for propagation
+    def plot_field(self, A_plot: np.ndarray) -> None:
+        """Plot a field for monitoring.
 
         Args:
-            A (np.ndarray): Array to transform.
-        Returns:
-            list: A list containing the FFT plans
+            A_plot (np.ndarray): Field to plot
         """
-        if self.backend == "GPU" and CUPY_AVAILABLE:
-            plan_fft = fftpack.get_fft_plan(A, axes=-1, value_type="C2C")
-            return [plan_fft]
-        else:
-            # try to load previous fftw wisdom
-            try:
-                with open("fft.wisdom", "rb") as file:
-                    wisdom = pickle.load(file)
-                    pyfftw.import_wisdom(wisdom)
-            except FileNotFoundError:
-                print("No FFT wisdom found, starting over ...")
-            plan_fft = pyfftw.FFTW(
-                A,
-                A,
-                direction="FFTW_FORWARD",
-                threads=multiprocessing.cpu_count(),
-                axes=(-1,),
-            )
-            plan_ifft = pyfftw.FFTW(
-                A,
-                A,
-                direction="FFTW_BACKWARD",
-                threads=multiprocessing.cpu_count(),
-                axes=(-1,),
-            )
-            with open("fft.wisdom", "wb") as file:
-                wisdom = pyfftw.export_wisdom()
-                pickle.dump(wisdom, file)
-            return [plan_fft, plan_ifft]
-
-    def split_step(
-        self,
-        A: np.ndarray,
-        V: np.ndarray,
-        propagator: np.ndarray,
-        plans: list,
-        precision: str = "single",
-    ) -> None:
-        """Split step function for one propagation step
-
-        Args:
-            A (np.ndarray): Field to propagate
-            V (np.ndarray): Potential field (can be None).
-            propagator (np.ndarray): Propagator matrix.
[…]
