+++ NLSE/nlse.py
@@ -875,6 +875,72 @@ class CNLSE(NLSE):
                 self._kernels.rabi_coupling(A1, A2, self.delta_z, self.omega / 2)
                 self._kernels.rabi_coupling(A2, A1_old, self.delta_z, self.omega / 2)
 
+    def _prepare_output_array(self, E: np.ndarray, normalize: bool) -> np.ndarray:
+        """Prepare the output array depending on backend."""
+        if self.backend == "GPU" and CUPY_AVAILABLE:
+            A = cp.empty_like(E)
+            A[:] = cp.asarray(E)
+            puiss_arr = cp.array([self.puiss, self.puiss2], dtype=E.dtype)
+        else:
+            A = pyfftw.empty_aligned(E.shape, dtype=E.dtype)
+            A[:] = E
+            puiss_arr = np.array([self.puiss, self.puiss2], dtype=E.dtype)
+        if normalize:
+            # normalization of the field
+            integral = (
+                (A.real * A.real + A.imag * A.imag) * self.delta_X * self.delta_Y
+            ).sum(axis=self._last_axes)
+            E_00 = (2 * puiss_arr / (c * epsilon_0 * integral)) ** 0.5
+            A = (E_00.T * A.T).T
+        return A
+
+    def plot_field(self, A_plot: np.ndarray) -> None:
+        """Plot the field.
+
+        Args:
+            A_plot (np.ndarray): The field to plot
+        """
+        # if array is multi-dimensional, drop dims until the shape is 2D
+        if A_plot.ndim > 3:
+            while len(A_plot.shape) > 3:
+                A_plot = A_plot[0]
+        if CUPY_AVAILABLE and isinstance(A_plot, cp.ndarray):
+            A_plot = A_plot.get()
+        fig = plt.figure(layout="constrained")
+        fig, ax = plt.subplots(2, 2, layout="constrained")
+        ext_real = [
+            self.X[0] * 1e3,
+            self.X[-1] * 1e3,
+            self.Y[0] * 1e3,
+            self.Y[-1] * 1e3,
+        ]
+        rho0 = np.abs(A_plot[0]) ** 2 * 1e-4 * c / 2 * epsilon_0
+        phi0 = np.angle(A_plot[0])
+        rho1 = np.abs(A_plot[1]) ** 2 * 1e-4 * c / 2 * epsilon_0
+        phi1 = np.angle(A_plot[1])
+        # plot amplitudes and phases
+        im0 = ax[0, 0].imshow(rho0, extent=ext_real)
+        ax[0, 0].set_title(r"$|\psi_1|^2$")
+        ax[0, 0].set_xlabel("x (mm)")
+        ax[0, 0].set_ylabel("y (mm)")
+        fig.colorbar(im0, ax=ax[0, 0], shrink=0.6, label="Intensity (W/cm^2)")
+        im1 = ax[0, 1].imshow(phi0, extent=ext_real, cmap="twilight_shifted")
+        ax[0, 1].set_title(r"Phase $\mathrm{arg}(\psi_1)$")
+        ax[0, 1].set_xlabel("x (mm)")
+        ax[0, 1].set_ylabel("y (mm)")
+        fig.colorbar(im1, ax=ax[0, 1], shrink=0.6, label="Phase (rad)")
+        im2 = ax[1, 0].imshow(rho1, extent=ext_real)
+        ax[1, 0].set_title(r"$|\psi_2|^2$")
+        ax[1, 0].set_xlabel("x (mm)")
+        ax[1, 0].set_ylabel("y (mm)")
[…]
