### 2024-03-27 b07439f
Big inheritance refactor
