+++ NLSE/nlse.py
@@ -135,7 +135,7 @@ class NLSE:
         else:
             self.nl_profile = np.ones((self.NY, self.NX), dtype=np.float32)
 
-    def build_propagator(self, k: float) -> np.ndarray:
+    def _build_propagator(self, k: float) -> np.ndarray:
         """Builds the linear propagation matrix
 
         Args:
@@ -148,7 +148,7 @@ class NLSE:
         propagator = np.exp(-1j * 0.5 * (self.Kxx**2 + self.Kyy**2) / k * self.delta_z)
         return propagator
 
-    def build_fft_plan(self, A: np.ndarray) -> list:
+    def _build_fft_plan(self, A: np.ndarray) -> list:
         """Builds the FFT plan objects for propagation
 
         Args:
@@ -189,6 +189,62 @@ class NLSE:
                 wisdom = pyfftw.export_wisdom()
                 pickle.dump(wisdom, file)
             return [plan_fft, plan_ifft]
+
+    def _prepare_output_array(self, E_in: np.ndarray, normalize: bool) -> np.ndarray:
+        """Prepare the output array depending on backend.
+
+        Args:
+            E_in (np.ndarray): Input array
+            normalize (bool): Normalize the field to the total power.
+        Returns:
+            np.ndarray: Output array
+        """
+        if self.backend == "GPU" and CUPY_AVAILABLE:
+            A = cp.empty_like(E_in)
+            A[:] = cp.asarray(E_in)
+        else:
+            A = pyfftw.empty_aligned(E_in.shape, dtype=E_in.dtype)
+            A[:] = E_in
+        if normalize:
+            # normalization of the field
+            integral = (
+                (A.real * A.real + A.imag * A.imag) * self.delta_X * self.delta_Y
+            ).sum(axis=self._last_axes)
+            E_00 = (2 * self.puiss / (c * epsilon_0 * integral)) ** 0.5
+            A = (E_00.T * A.T).T
+        return A
+
+    def _send_arrays_to_gpu(self) -> None:
+        """
+        Send arrays to GPU.
+        """
+        if self.V is not None:
+            self.V = cp.asarray(self.V)
+        self.nl_profile = cp.asarray(self.nl_profile)
+        self.propagator = cp.asarray(self.propagator)
+        # for broadcasting of parameters in case they are
+        # not already on the GPU
+        if isinstance(self.n2, np.ndarray):
+            self.n2 = cp.asarray(self.n2)
+        if isinstance(self.alpha, np.ndarray):
+            self.alpha = cp.asarray(self.alpha)
+        if isinstance(self.I_sat, np.ndarray):
+            self.I_sat = cp.asarray(self.I_sat)
+
+    def _retrieve_arrays_from_gpu(self) -> None:
+        """
+        Retrieve arrays from GPU.
+        """
+        if self.V is not None:
+            self.V = self.V.get()
+        self.nl_profile = self.nl_profile.get()
+        self.propagator = self.propagator.get()
+        if isinstance(self.n2, cp.ndarray):
+            self.n2 = self.n2.get()
+        if isinstance(self.alpha, cp.ndarray):
+            self.alpha = self.alpha.get()
+        if isinstance(self.I_sat, cp.ndarray):
+            self.I_sat = self.I_sat.get()
 
     def split_step(
         self,
@@ -298,62 +354,6 @@ class NLSE:
                     2 * self.I_sat / (epsilon_0 * c),
                 )
 
-    def _prepare_output_array(self, E_in: np.ndarray, normalize: bool) -> np.ndarray:
-        """Prepare the output array depending on backend.
-
-        Args:
-            E_in (np.ndarray): Input array
-            normalize (bool): Normalize the field to the total power.
-        Returns:
[…]
