### 2024-03-27 f284a01
Non locality in CNLSE_1d
