### 2024-03-27 4859bd3
Inheritance of GPE
