+++ NLSE/nlse.py
@@ -139,9 +139,7 @@ class NLSE:
         """Builds the linear propagation matrix
 
         Args:
-            precision (str, optional): "single" or "double" application of the
-            propagator.
-            Defaults to "single".
+            k (float): Wavenumber
         Returns:
             propagator (np.ndarray): the propagator matrix
         """
@@ -544,9 +542,7 @@ class NLSE_1d(NLSE):
         """Builds the linear propagation matrix
 
         Args:
-            precision (str, optional): "single" or "double" application of the
-            propagator.
-            Defaults to "single".
+            k (float): Wavenumber
         Returns:
             propagator (np.ndarray): the propagator matrix
         """
@@ -1217,8 +1213,7 @@ class CNLSE_1d(CNLSE):
         plt.show()
 
 
-# TODO(Tangui): Setup inheritance from NLSE
-class GPE:
+class GPE(NLSE):
     """A class to solve GPE."""
 
     def __init__(
@@ -1253,387 +1248,111 @@ class GPE:
             nl_length (float, optional): Non local length scale in m. Defaults to 0.
             backend (str, optional): "GPU" or "CPU". Defaults to BACKEND.
         """
-        self.backend = backend
-        if self.backend == "GPU" and CUPY_AVAILABLE:
-            self._kernels = kernels_gpu
-            self._convolution = signal_cp.fftconvolve
-        else:
-            self._kernels = kernels_cpu
-            self._convolution = signal.fftconvolve
+        super().__init__(
+            alpha=gamma,
+            puiss=N,
+            window=window,
+            n2=g,
+            V=V,
+            L=0,
+            NX=NX,
+            NY=NY,
+            Isat=sat,
+            nl_length=nl_length,
+            wvl=2 * np.pi / m,
+            backend=backend,
+        )
         # listof physical parameters
         self.g = g
         self.V = V
-        self.gamma = gamma
-        self.N = N
-        self.m = m
+        self.gamma = self.alpha
+        self.N = self.puiss
+        self.m = self.k
         self.sat = sat
-        # number of grid points in X (even, best is power of 2 or
-        # low prime factors)
-        self.NX = NX
-        self.NY = NY
-        self.window = window
-        tau_nl = self.window**2 / (abs(self.g) * self.N)
-        self.delta_t = tau_nl / 10
-        # transverse coordinate
-        self.X, self.delta_X = np.linspace(
-            -self.window / 2,
-            self.window / 2,
-            num=NX,
-            endpoint=False,
-            retstep=True,
-            dtype=np.float32,
-        )
-        self.Y, self.delta_Y = np.linspace(
-            -self.window / 2,
-            self.window / 2,
-            num=NY,
-            endpoint=False,
-            retstep=True,
-            dtype=np.float32,
-        )
+        self.delta_t = self.delta_z
+        # do some conversion for the units
+        self.I_sat *= epsilon_0 * c / 2
 
-        self.XX, self.YY = np.meshgrid(self.X, self.Y)
-        # definition of the Fourier frequencies for the linear step
-        self.Kx = 2 * np.pi * np.fft.fftfreq(self.NX, d=self.delta_X)
-        self.Ky = 2 * np.pi * np.fft.fftfreq(self.NY, d=self.delta_Y)
-        self.Kxx, self.Kyy = np.meshgrid(self.Kx, self.Ky)
-        self.propagator = None
-        self.plans = None
-        self.nl_length = nl_length
-        if self.nl_length > 0:
-            d = self.nl_length // self.delta_X
-            x = np.arange(-3 * d, 3 * d + 1)
-            y = np.arange(-3 * d, 3 * d + 1)
-            XX, YY = np.meshgrid(x, y)
-            R = np.hypot(XX, YY)
[…]
