+++ NLSE/cnlse.py
@@ -1,0 +1,508 @@
+from .nlse import NLSE
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy.constants import c, epsilon_0
+import pyfftw
+import time
+import tqdm
+from .utils import __BACKEND__, __CUPY_AVAILABLE__
+
+if __CUPY_AVAILABLE__:
+    import cupy as cp
+
+
+class CNLSE(NLSE):
+    """A class to solve the coupled NLSE"""
+
+    def __init__(
+        self,
+        alpha: float,
+        puiss: float,
+        window: float,
+        n2: float,
+        n12: float,
+        V: np.ndarray,
+        L: float,
+        NX: int = 1024,
+        NY: int = 1024,
+        Isat: float = np.inf,
+        nl_length: float = 0,
+        wvl: float = 780e-9,
+        omega: float = None,
+        backend: str = __BACKEND__,
+    ) -> object:
+        """Instantiates the class with all the relevant physical parameters
+
+        Args:
+            alpha (float): alpha through the cell
+            puiss (float): Optical power in W
+            waist (float): Beam waist in m
+            window (float): Computational window in m
+            n2 (float): Non linear index of the 1 st component in m^2/W
+            n12 (float): Inter component interaction parameter
+            V (np.ndarray): Potential landscape in a.u
+            L (float): Length of the cell in m
+            NX (int, optional): Number of points along x. Defaults to 1024.
+            NY (int, optional): Number of points along y. Defaults to 1024.
+            Isat (float, optional): Saturation intensity, assumed to be the same
+            for both components. Defaults to infinity.
+            nl_length (float, optional): Nonlocal length. Defaults to 0.
+            wvl (float, optional): Wavelength in m. Defaults to 780 nm.
+            omega (float, optional): Rabi coupling. Defaults to None.
+            __BACKEND__ (str, optional): "GPU" or "CPU". Defaults to __BACKEND__.
+        Returns:
+            object: CNLSE class instance
+        """
+        super().__init__(
+            alpha=alpha,
+            puiss=puiss,
+            window=window,
+            n2=n2,
+            V=V,
+            L=L,
+            NX=NX,
+            NY=NY,
+            Isat=Isat,
+            nl_length=nl_length,
+            wvl=wvl,
+            backend=backend,
+        )
+        self.n12 = n12
+        self.I_sat2 = self.I_sat
+        # initialize intra component 2 interaction parameter
+        # to be the same as intra component 1
+        self.n22 = self.n2
+        # Rabi coupling
+        self.omega = omega
+        # same for the losses, this is to leave separate attributes so
+        # the the user can chose whether or not to unbalence the rates
+        self.alpha2 = self.alpha
+        # wavenumbers
+        self.k2 = self.k
+        # powers
+        self.puiss2 = self.puiss
+        # waists
+        self.propagator1 = None
+        self.propagator2 = None
+
+    def _prepare_output_array(self, E: np.ndarray, normalize: bool) -> np.ndarray:
+        """Prepare the output array depending on __BACKEND__."""
+        if self.backend == "GPU" and __CUPY_AVAILABLE__:
+            A = cp.empty_like(E)
+            A[:] = cp.asarray(E)
+            puiss_arr = cp.array([self.puiss, self.puiss2], dtype=E.dtype)
+        else:
+            A = pyfftw.empty_aligned(E.shape, dtype=E.dtype)
+            A[:] = E
+            puiss_arr = np.array([self.puiss, self.puiss2], dtype=E.dtype)
+        if normalize:
[…]
