+++ NLSE/nlse.py
@@ -10,31 +10,21 @@ import tqdm
 import matplotlib.pyplot as plt
 import numpy as np
 import pyfftw
-from scipy.constants import c, epsilon_0, hbar, atomic_mass
+from scipy.constants import c, epsilon_0
 from scipy import signal
 from scipy import special
 from . import kernels_cpu
+from .utils import __BACKEND__, __CUPY_AVAILABLE__
 
-BACKEND = "GPU"
-
-if BACKEND == "GPU":
-    try:
-        import cupy as cp
-        import cupyx.scipy.fftpack as fftpack
-        import cupyx.scipy.signal as signal_cp
-
-        CUPY_AVAILABLE = True
-        from . import kernels_gpu
-
-    except ImportError:
-        print("CuPy not available, falling back to CPU backend ...")
-        CUPY_AVAILABLE = False
-        BACKEND = "CPU"
+if __CUPY_AVAILABLE__:
+    import cupy as cp
+    import cupyx.scipy.fftpack as fftpack
+    import cupyx.scipy.signal as signal_cp
+    from . import kernels_gpu
 
 pyfftw.config.NUM_THREADS = multiprocessing.cpu_count()
 pyfftw.config.PLANNER_EFFORT = "FFTW_MEASURE"
 pyfftw.interfaces.cache.enable()
-np.random.seed(1)
 
 
 class NLSE:
@@ -53,7 +43,7 @@ class NLSE:
         Isat: float = np.inf,
         nl_length: float = 0,
         wvl: float = 780e-9,
-        backend: str = BACKEND,
+        backend: str = __BACKEND__,
     ) -> object:
         """Instantiates the simulation.
         Solves an equation : d/dz psi = -1/2k0(d2/dx2 + d2/dy2) psi + k0 dn psi +
@@ -67,11 +57,12 @@ class NLSE:
             Isat (float): Saturation intensity in W/m^2
             nl_length (float): Non linear length in m
             wvl (float): Wavelength in m
-            backend (str, optional): "GPU" or "CPU". Defaults to BACKEND.
+            __BACKEND__ (str, optional): "GPU" or "CPU". Defaults to __BACKEND__.
         """
         # listof physical parameters
+        global __CUPY_AVAILABLE__
         self.backend = backend
-        if self.backend == "GPU" and CUPY_AVAILABLE:
+        if self.backend == "GPU" and __CUPY_AVAILABLE__:
             self._kernels = kernels_gpu
             self._convolution = signal_cp.fftconvolve
         else:
@@ -154,7 +145,7 @@ class NLSE:
         Returns:
             list: A list containing the FFT plans
         """
-        if self.backend == "GPU" and CUPY_AVAILABLE:
+        if self.backend == "GPU" and __CUPY_AVAILABLE__:
             plan_fft = fftpack.get_fft_plan(
                 A,
                 axes=self._last_axes,
@@ -189,7 +180,7 @@ class NLSE:
             return [plan_fft, plan_ifft]
 
     def _prepare_output_array(self, E_in: np.ndarray, normalize: bool) -> np.ndarray:
-        """Prepare the output array depending on backend.
+        """Prepare the output array depending on __BACKEND__.
 
         Args:
             E_in (np.ndarray): Input array
@@ -197,7 +188,7 @@ class NLSE:
         Returns:
             np.ndarray: Output array
         """
-        if self.backend == "GPU" and CUPY_AVAILABLE:
+        if self.backend == "GPU" and __CUPY_AVAILABLE__:
             A = cp.empty_like(E_in)
             A[:] = cp.asarray(E_in)
         else:
@@ -265,7 +256,7 @@ class NLSE:
             propagation step.
             Defaults to "single".
         """
[…]
