+++ NLSE/__init__.py
@@ -1,5 +1,6 @@
+from . import utils
 from .nlse import NLSE
-from .nlse import NLSE_1d
-from .nlse import CNLSE
-from .nlse import CNLSE_1d
-from .nlse import GPE
+from .nlse_1d import NLSE_1d
+from .cnlse import CNLSE
+from .cnlse_1d import CNLSE_1d
+from .gpe import GPE
