### 2024-03-28 0ca858a
Pass __CUPY_AVAILABLE__ as default attribute
