+++ tests/test_nlse.py
@@ -1,5 +1,6 @@
 from NLSE import NLSE
 import numpy as np
+import pyfftw
 
 if NLSE.__CUPY_AVAILABLE__:
     import cupy as cp
@@ -24,9 +25,16 @@ def test_build_propagator() -> None:
             alpha, puiss, window, n2, None, L, NX=N, NY=N, Isat=Isat, backend=backend
         )
         prop = simu._build_propagator(simu.k)
-        assert prop == np.exp(
-            -1j * 0.5 * (simu.Kxx**2 + simu.Kyy**2) / simu.k * simu.delta_z
-        )
+        if backend == "CPU":
+            assert np.allclose(
+                prop,
+                np.exp(-1j * 0.5 * (simu.Kxx**2 + simu.Kyy**2) / simu.k * simu.delta_z),
+            )
+        else:
+            assert cp.allclose(
+                prop,
+                np.exp(-1j * 0.5 * (simu.Kxx**2 + simu.Kyy**2) / simu.k * simu.delta_z),
+            )
 
 
 def test_build_fft_plan() -> None:
@@ -38,14 +46,64 @@ def test_build_fft_plan() -> None:
             A = np.random.random((N, N)) + 1j * np.random.random((N, N))
         elif backend == "GPU" and NLSE.__CUPY_AVAILABLE__:
             A = cp.random.random((N, N)) + 1j * cp.random.random((N, N))
+        plans = simu._build_fft_plan(A)
+        if backend == "CPU":
+            assert len(plans) == 2
+            assert isinstance(plans[0], pyfftw.FFTW)
+            assert plans[0].output_shape == (N, N)
+        else:
+            assert len(plans) == 1
+            assert isinstance(plans[0], cp.cuda.cufft.PlanNd)
+            assert plans[0].shape == (N, N)
 
 
 def test_prepare_output_array() -> None:
-    pass
+    for backend in ["CPU", "GPU"]:
+        simu = NLSE(
+            alpha, puiss, window, n2, None, L, NX=N, NY=N, Isat=Isat, backend=backend
+        )
+        if backend == "CPU":
+            A = np.random.random((N, N)) + 1j * np.random.random((N, N))
+        elif backend == "GPU" and NLSE.__CUPY_AVAILABLE__:
+            A = cp.random.random((N, N)) + 1j * cp.random.random((N, N))
+        out = simu._prepare_output_array(A, normalize=True)
+        if backend == "CPU":
+            assert isinstance(out, np.ndarray)
+            out /= np.max(np.abs(out))
+            A /= np.max(np.abs(A))
+            assert np.allclose(out, A)
+        else:
+            assert isinstance(out, cp.ndarray)
+            assert out.shape == (N, N)
+            out /= cp.max(cp.abs(out))
+            A /= cp.max(cp.abs(A))
+            assert cp.allclose(out, A)
 
 
 def test_send_arrays_to_gpu() -> None:
-    pass
+    if NLSE.__CUPY_AVAILABLE__:
+        global alpha
+        global n2
+        global Isat
+        V = np.random.random((N, N)) + 1j * np.random.random((N, N))
+        alpha = np.repeat(alpha, 2)
+        alpha = alpha[..., cp.newaxis, cp.newaxis]
+        n2 = np.repeat(n2, 2)
+        n2 = n2[..., cp.newaxis, cp.newaxis]
+        Isat = np.repeat(Isat, 2)
+        Isat = Isat[..., cp.newaxis, cp.newaxis]
+        simu = NLSE(
+            alpha, puiss, window, n2, V, L, NX=N, NY=N, Isat=Isat, backend="GPU"
+        )
[…]
