### 2024-03-28 3f8b63e
Tests for NLSE
