### 2024-03-28 7f18dba
Correct  check with cupy available
