### 2024-03-29 a0f3190
tests NLSE / NLSE_1d / GPE
