+++ tests/run_all_tests.py
@@ -1,4 +1,4 @@
-import test_cnles_1d
+import tests.test_cnlse_1d as test_cnlse_1d
 import test_cnlse
 import test_nlse_1d
 import test_nlse
@@ -6,7 +6,7 @@ import test_gpe
 import test_nonlocality
 
 if __name__ == "__main__":
-    test_cnles_1d.main()
+    test_cnlse_1d.main()
     test_cnlse.main()
     test_nlse_1d.main()
     test_nlse.main()
