+++ tests/test_cnlse.py
@@ -1,5 +1,9 @@
 from NLSE import CNLSE
 import numpy as np
+from scipy.constants import c, epsilon_0
+
+if CNLSE.__CUPY_AVAILABLE__:
+    import cupy as cp
 
 PRECISION_COMPLEX = np.complex64
 PRECISION_REAL = np.float32
@@ -14,6 +18,229 @@ puiss = 1.05
 Isat = 10e4  # saturation intensity in W/m^2
 L = 1e-3
 alpha = 20
+
+
+def test_prepare_output_array() -> None:
+    for backend in ["CPU", "GPU"]:
+        simu = CNLSE(
+            alpha,
+            puiss,
+            window,
+            n2,
+            n12,
+            None,
+            L,
+            NX=N,
+            NY=N,
+            Isat=Isat,
+            backend=backend,
+        )
+        if backend == "CPU":
+            A = np.ones((2, N, N), dtype=PRECISION_COMPLEX)
+        elif backend == "GPU" and CNLSE.__CUPY_AVAILABLE__:
+            A = cp.ones((2, N, N), dtype=PRECISION_COMPLEX)
+        out = simu._prepare_output_array(A, normalize=True)
+        integral = (
+            (out.real * out.real + out.imag * out.imag) * simu.delta_X * simu.delta_Y
+        ).sum(axis=simu._last_axes)
+        integral *= c * epsilon_0 / 2
+        assert np.allclose(
+            integral,
+            np.array([simu.puiss, simu.puiss2]),
+            rtol=1e-4,
+        ), "Normalization failed."
+        assert out.shape == (2, N, N), "Output array has wrong shape."
+        if backend == "CPU":
+            assert isinstance(
+                out, np.ndarray
+            ), "Ouptut array type does not match backend."
+            out /= np.max(np.abs(out))
+            A /= np.max(np.abs(A))
+            assert np.allclose(out, A), "Output array does not match input array."
+        elif backend == "GPU" and CNLSE.__CUPY_AVAILABLE__:
+            assert isinstance(
+                out, cp.ndarray
+            ), "Ouptut array type does not match backend."
+            out /= cp.max(cp.abs(out))
+            A /= cp.max(cp.abs(A))
+            assert cp.allclose(out, A), "Output array does not match input array."
+
+
+def test_send_arrays_to_gpu() -> None:
+    if CNLSE.__CUPY_AVAILABLE__:
+        alpha = 20
+        Isat = 10e4
+        n2 = -1.6e-9
+        n12 = -1e-10
+        V = np.random.random((N, N)) + 1j * np.random.random((N, N))
+        alpha = np.repeat(alpha, 2)
+        alpha = alpha[..., np.newaxis, np.newaxis, np.newaxis]
+        n2 = np.repeat(n2, 2)
+        n2 = n2[..., np.newaxis, np.newaxis, np.newaxis]
+        n12 = np.repeat(n2, 2)
+        n12 = n12[..., np.newaxis, np.newaxis, np.newaxis]
+        Isat = np.repeat(Isat, 2)
+        Isat = Isat[..., np.newaxis, np.newaxis, np.newaxis]
+        simu = CNLSE(
+            alpha,
+            puiss,
+            window,
+            n2,
+            n12,
+            V,
+            L,
+            NX=N,
+            NY=N,
+            Isat=Isat,
+            backend="GPU",
+        )
+        simu.propagator1 = simu._build_propagator(simu.k)
+        simu.propagator2 = simu._build_propagator(simu.k2)
+        simu._send_arrays_to_gpu()
+        assert isinstance(
+            simu.propagator1, cp.ndarray
+        ), "propagator1 is not a cp.ndarray."
[…]
