+++ tests/test_cnlse_1d.py
@@ -1,0 +1,126 @@
+from NLSE import CNLSE_1d
+import numpy as np
+from scipy.constants import c, epsilon_0
+
+if CNLSE_1d.__CUPY_AVAILABLE__:
+    import cupy as cp
+PRECISION_COMPLEX = np.complex64
+PRECISION_REAL = np.float32
+N = 2048
+n2 = -1.6e-9
+n12 = -1e-10
+waist = 2.23e-3
+waist2 = 70e-6
+window = 4 * waist
+puiss = 1.05
+Isat = 10e4  # saturation intensity in W/m^2
+L = 1e-3
+alpha = 20
+
+
+def test_build_propagator() -> None:
+    for backend in ["CPU", "GPU"]:
+        simu = CNLSE_1d(
+            alpha, puiss, window, n2, n12, None, L, NX=N, Isat=Isat, backend=backend
+        )
+        prop = simu._build_propagator(simu.k)
+        assert np.allclose(
+            prop, np.exp(-1j * 0.5 * (simu.Kx**2) / simu.k * simu.delta_z)
+        ), "Propagator is wrong."
+
+
+def test_prepare_output_array() -> None:
+    for backend in ["CPU", "GPU"]:
+        simu = CNLSE_1d(
+            alpha,
+            puiss,
+            window,
+            n2,
+            n12,
+            None,
+            L,
+            NX=N,
+            Isat=Isat,
+            backend=backend,
+        )
+        if backend == "CPU":
+            A = np.ones((2, N), dtype=PRECISION_COMPLEX)
+        elif backend == "GPU" and CNLSE_1d.__CUPY_AVAILABLE__:
+            A = cp.ones((2, N), dtype=PRECISION_COMPLEX)
+        out = simu._prepare_output_array(A, normalize=True)
+        integral = ((out.real * out.real + out.imag * out.imag) * simu.delta_X).sum(
+            axis=simu._last_axes
+        ) ** 2
+        integral *= c * epsilon_0 / 2
+        assert np.allclose(
+            integral,
+            np.array([simu.puiss, simu.puiss2]),
+            rtol=1e-4,
+        ), "Normalization failed."
+        assert out.shape == (2, N), "Output array has wrong shape."
+        if backend == "CPU":
+            assert isinstance(
+                out, np.ndarray
+            ), "Ouptut array type does not match backend."
+            out /= np.max(np.abs(out))
+            A /= np.max(np.abs(A))
+            assert np.allclose(out, A), "Output array does not match input array."
+        elif backend == "GPU" and CNLSE_1d.__CUPY_AVAILABLE__:
+            assert isinstance(
+                out, cp.ndarray
+            ), "Ouptut array type does not match backend."
+            out /= cp.max(cp.abs(out))
+            A /= cp.max(cp.abs(A))
+            assert cp.allclose(out, A), "Output array does not match input array."
+
+
+def test_out_field() -> None:
+    for backend in ["CPU", "GPU"]:
+        simu = CNLSE_1d(
+            0, puiss, window, n2, n12, None, L, NX=N, Isat=Isat, backend=backend
+        )
+        E0 = np.ones((2, N), dtype=PRECISION_COMPLEX)
+        A = simu.out_field(E0, L, verbose=False, plot=False, precision="single")
+        integral = ((A.real * A.real + A.imag * A.imag) * simu.delta_X).sum(
+            axis=simu._last_axes
+        )
+        integral *= c * epsilon_0 / 2
+        assert A.shape == (2, N), "Output array has wrong shape."
+        assert np.allclose(
+            integral, [simu.puiss, simu.puiss2], rtol=1e-4
+        ), "Normalization failed."
+
+
[…]
