### 2024-03-29 addb2bb
Optional deps and github test workflow
