### 2024-03-29 #6509 @taladjidi
@RMeli I significantly refactored the code:
- I split all the classes into separate files, using inheritance as much as possible to limit code duplication
- Added functions in order to make the logic more legible
- Wrote a series of tests using Pytest, with the corresponding GitHub action
At the moment though, one test fails inexplicably in the GitHub action, while it passes on all the computers I have access to in the lab (Windows, MacOs or Linux). I did not manage to diagnosticate the issue for now ...
I hope this adresses most of your remarks 😀
