+++ tests/test_nlse.py
@@ -29,7 +29,7 @@ def test_build_propagator() -> None:
         assert np.allclose(
             prop,
             np.exp(-1j * 0.5 * (simu.Kxx**2 + simu.Kyy**2) / simu.k * simu.delta_z),
-        ), "Propagator is wrong."
+        ), f"Propagator is wrong. (Backend {backend})"
 
 
 def test_build_fft_plan() -> None:
@@ -43,13 +43,20 @@ def test_build_fft_plan() -> None:
             A = cp.random.random((N, N)) + 1j * cp.random.random((N, N))
         plans = simu._build_fft_plan(A)
         if backend == "CPU":
-            assert len(plans) == 2, "Number of plans is wrong."
-            assert isinstance(plans[0], pyfftw.FFTW), "Plan type is wrong."
-            assert plans[0].output_shape == (N, N), "Plan shape is wrong."
+            assert len(plans) == 2, f"Number of plans is wrong. (Backend {backend})"
+            assert isinstance(
+                plans[0], pyfftw.FFTW
+            ), f"Plan type is wrong. (Backend {backend})"
+            assert plans[0].output_shape == (
+                N,
+                N,
+            ), f"Plan shape is wrong. (Backend {backend})"
         elif backend == "GPU" and NLSE.__CUPY_AVAILABLE__:
-            assert len(plans) == 1, "Number of plans is wrong."
-            assert isinstance(plans[0], cp.cuda.cufft.PlanNd), "Plan type is wrong."
-            assert plans[0].shape == (N, N), "Plan shape is wrong."
+            assert len(plans) == 1, f"Number of plans is wrong. (Backend {backend})"
+            assert isinstance(
+                plans[0], cp.cuda.cufft.PlanNd
+            ), f"Plan type is wrong. (Backend {backend})"
+            assert plans[0].shape == (N, N), f"Plan shape is wrong. (Backend {backend})"
 
 
 def test_prepare_output_array() -> None:
@@ -66,22 +73,28 @@ def test_prepare_output_array() -> None:
             (out.real * out.real + out.imag * out.imag) * simu.delta_X * simu.delta_Y
         ).sum(axis=simu._last_axes)
         integral *= c * epsilon_0 / 2
-        assert np.allclose(integral, simu.puiss), "Normalization failed."
-        assert out.shape == (N, N), "Output array has wrong shape."
+        assert np.allclose(
+            integral, simu.puiss
+        ), f"Normalization failed. (Backend {backend})"
+        assert out.shape == (N, N), f"Output array has wrong shape. (Backend {backend})"
         if backend == "CPU":
             assert isinstance(
                 out, np.ndarray
-            ), "Output array type does not match backend."
+            ), f"Output array type does not match backend. (Backend {backend})"
             out /= np.max(np.abs(out))
             A /= np.max(np.abs(A))
-            assert np.allclose(out, A), "Output array does not match input array."
+            assert np.allclose(
+                out, A
+            ), f"Output array does not match input array. (Backend {backend})"
         elif backend == "GPU" and NLSE.__CUPY_AVAILABLE__:
             assert isinstance(
                 out, cp.ndarray
-            ), "Output array type does not match backend."
+            ), f"Output array type does not match backend. (Backend {backend})"
             out /= cp.max(cp.abs(out))
             A /= cp.max(cp.abs(A))
-            assert cp.allclose(out, A), "Output array does not match input array."
+            assert cp.allclose(
+                out, A
[…]
