### 2024-04-02 c18927d
Start of NLSE_3d, some docstring correction
