### 2024-04-03 21a3dd3
Memory alignment issues that was breaking FFT planning
