+++ NLSE/gpe.py
@@ -77,7 +77,7 @@ class GPE(NLSE):
         """
         propagator = np.exp(
             -1j * 0.5 * hbar * (self.Kxx**2 + self.Kyy**2) / m * self.delta_t
-        )
+        ).astype(np.complex64)
         return propagator
 
     def _prepare_output_array(self, E_in: np.ndarray, normalize: bool) -> np.ndarray:
