+++ __init__.py
@@ -5,7 +5,7 @@ A package for solving the Nonlinear Schrödinger Equation (NLSE) using the
 Split-Step Fourier method.
 """
 
-__version__ = "2.0.0"
+__version__ = "2.0.2"
 __author__ = "Tangui Aladjidi"
 __license__ = "MIT"
 __credits__ = "Laboratoire Kastler Brossel, Paris, France"
