### 2024-04-04 f5aefe9
Propagator simplification
