+++ NLSE/cnlse.py
@@ -108,39 +108,30 @@ class CNLSE(NLSE):
         """
         Send arrays to GPU.
         """
-        if self.V is not None:
-            self.V = cp.asarray(self.V)
-        self.nl_profile = cp.asarray(self.nl_profile)
-        self.propagator1 = cp.asarray(self.propagator1)
-        self.propagator2 = cp.asarray(self.propagator2)
+        super()._send_arrays_to_gpu()
         # for broadcasting of parameters in case they are
         # not already on the GPU
-        if isinstance(self.n2, np.ndarray):
-            self.n2 = cp.asarray(self.n2)
+        if isinstance(self.n22, np.ndarray):
+            self.n22 = cp.asarray(self.n22)
         if isinstance(self.n12, np.ndarray):
             self.n12 = cp.asarray(self.n12)
-        if isinstance(self.alpha, np.ndarray):
-            self.alpha = cp.asarray(self.alpha)
-        if isinstance(self.I_sat, np.ndarray):
-            self.I_sat = cp.asarray(self.I_sat)
 
     def _retrieve_arrays_from_gpu(self) -> None:
         """
         Retrieve arrays from GPU.
         """
-        if self.V is not None:
-            self.V = self.V.get()
-        self.nl_profile = self.nl_profile.get()
-        self.propagator1 = self.propagator1.get()
-        self.propagator2 = self.propagator2.get()
+        super()._retrieve_arrays_from_gpu()
         if isinstance(self.n2, cp.ndarray):
-            self.n2 = self.n2.get()
+            self.n22 = self.n22.get()
         if isinstance(self.n12, cp.ndarray):
             self.n12 = self.n12.get()
-        if isinstance(self.alpha, cp.ndarray):
-            self.alpha = self.alpha.get()
-        if isinstance(self.I_sat, cp.ndarray):
-            self.I_sat = self.I_sat.get()
+
+    def _build_propagator(self) -> np.ndarray:
+        propagator1 = super()._build_propagator()
+        propagator2 = np.exp(
+            -1j * 0.5 * (self.Kxx**2 + self.Kyy**2) / self.k2 * self.delta_z
+        ).astype(np.complex64)
+        return np.array([propagator1, propagator2])
 
     def _take_components(self, A: np.ndarray) -> tuple:
         """Take the components of the field.
@@ -158,8 +149,7 @@ class CNLSE(NLSE):
         self,
         A: np.ndarray,
         V: np.ndarray,
-        propagator1: np.ndarray,
-        propagator2: np.ndarray,
+        propagator: np.ndarray,
         plans: list,
         precision: str = "single",
     ) -> None:
@@ -249,17 +239,14 @@ class CNLSE(NLSE):
         if self.backend == "GPU" and self.__CUPY_AVAILABLE__:
             plan_fft.fft(A, A, cp.cuda.cufft.CUFFT_FORWARD)
             # linear step in Fourier domain (shifted)
-            cp.multiply(A1, propagator1, out=A1)
-            cp.multiply(A2, propagator2, out=A2)
-            # A = cp.fft.ifftn(A, axes=self._last_axes)
+            cp.multiply(A, propagator, out=A)
             plan_fft.fft(A, A, cp.cuda.cufft.CUFFT_INVERSE)
             # fft normalization
             A *= 1 / np.prod(A.shape[self._last_axes[0] :])
         else:
             plan_fft, plan_ifft = plans
             plan_fft(input_array=A, output_array=A)
-            np.multiply(A1, propagator1, out=A1)
-            np.multiply(A2, propagator2, out=A2)
+            np.multiply(A, propagator, out=A)
[…]
