### 2024-04-04 44144bb
propagator simplification
