### 2024-04-04 810a3a9
plot for nlse3d
