### 2024-04-04 69beb77
basic infra for nlse_3d
