+++ NLSE/__init__.py
@@ -1,6 +1,7 @@
 from . import utils
 from .nlse import NLSE
 from .nlse_1d import NLSE_1d
+from .nlse_3d import NLSE_3d
 from .cnlse import CNLSE
 from .cnlse_1d import CNLSE_1d
 from .gpe import GPE
