+++ NLSE/nlse_3d.py
@@ -17,7 +17,7 @@ class NLSE_3d(NLSE):
     def __init__(
         self,
         alpha: float,
-        puiss: float,
+        energy: float,
         window: np.ndarray,
         n2: float,
         D0: float,
@@ -37,7 +37,7 @@ class NLSE_3d(NLSE):
           k0 n2 psi**2 psi
         Args:
             alpha (float): alpha
-            puiss (float): Power in W
+            energy (float): Total energy in J
             window (np.ndarray): Computanional window in the transverse plane (index 0) in m
             and longitudinal direction (index 1) in s.
             n2 (float): Non linear coeff in m^2/W.
@@ -55,7 +55,7 @@ class NLSE_3d(NLSE):
         """
         super().__init__(
             alpha=alpha,
-            puiss=puiss,
+            puiss=energy,
             window=window[0],
             n2=n2,
             V=V,
@@ -67,6 +67,7 @@ class NLSE_3d(NLSE):
             wvl=wvl,
             backend=backend,
         )
+        self.energy = self.puiss
         self.NZ = NZ
         self.window_t = window[1]
         self.T, self.delta_T = np.linspace(
@@ -75,6 +76,7 @@ class NLSE_3d(NLSE):
         self.omega = 2 * np.pi * np.fft.fftfreq(self.NZ, self.delta_T)
         self.D0 = D0
         self.vg = vg
+        self.XX, self.YY, self.TT = np.meshgrid(self.X, self.Y, self.T)
         self.Kxx, self.Kyy, self.Omega = np.meshgrid(self.Kx, self.Ky, self.omega)
         self._last_axes = (-3, -2, -1)  # Axes are x, y, t
 
@@ -83,6 +85,35 @@ class NLSE_3d(NLSE):
         prop_t = np.exp(-1j * self.D0 / 2 * self.Omega**2)
         # prop_t *= np.exp(1 / self.vg * self.Omega)
         return prop_2d * prop_t
+
+    def _prepare_output_array(self, E_in: np.ndarray, normalize: bool) -> np.ndarray:
+        """Prepare the output array depending on __BACKEND__.
+
+        Args:
+            E_in (np.ndarray): Input array
+            normalize (bool): Normalize the field to the total power.
+        Returns:
+            np.ndarray: Output array
+        """
+        if self.backend == "GPU" and self.__CUPY_AVAILABLE__:
+            A = cp.empty_like(E_in)
+            E_in = cp.asarray(E_in)
+        else:
+            A = pyfftw.empty_aligned(
+                E_in.shape, dtype=E_in.dtype, n=pyfftw.simd_alignment
+            )
+        if normalize:
+            # normalization of the field
+            integral = (
+                (E_in.real * E_in.real + E_in.imag * E_in.imag)
+                * self.delta_X
+                * self.delta_Y
+                * self.delta_T
+            ).sum(axis=self._last_axes)
+            integral *= c * epsilon_0 / 2
+            E_00 = (self.energy / integral) ** 0.5
+            A[:] = (E_00.T * E_in.T).T
+        return A
 
     def plot_field(self, A_plot: np.ndarray) -> None:
         # if array is multi-dimensional, drop dims until the shape is 2D
@@ -99,10 +130,10 @@ class NLSE_3d(NLSE):
[…]
