### 2024-04-04 7103ff7
faster convolve for non locality
