### 2024-04-04 #6509 @RMeli
Thanks @taladjidi. I'll start looking for additional reviewers.

> At the moment though, one test fails inexplicably in the GitHub action

Loos like `norm` ends up being `0.0`. Do you have any update on troubleshooting the issue?
