### 2024-04-05 #6509 @taladjidi
Hi, yes I saw. I did not manage to replicate the issue. I thought it might be a cache issue since I do a lot of jitting, but even when I clear all caches, I do not manage to replicate this zero norm problem (I created an environment cloning the packages versions used by GitHub). I'll spend some more energy on this today !
