+++ tests/test_nlse_1d.py
@@ -72,6 +72,31 @@ def test_prepare_output_array() -> None:
             ), f"Output array does not match input array. (Backend {backend})"
 
 
+def test_split_step() -> None:
+    for backend in ["CPU", "GPU"]:
+        simu = NLSE_1d(
+            alpha, puiss, window, n2, None, L, NX=N, Isat=Isat, backend=backend
+        )
+        simu.delta_z = 0
+        simu.propagator = simu._build_propagator()
+        E = np.ones((N,), dtype=PRECISION_COMPLEX)
+        A = simu._prepare_output_array(E, normalize=False)
+        simu.plans = simu._build_fft_plan(A)
+        simu.propagator = simu._build_propagator()
+        if backend == "GPU" and NLSE_1d.__CUPY_AVAILABLE__:
+            E = cp.asarray(E)
+            simu._send_arrays_to_gpu()
+        simu.split_step(E, simu.V, simu.propagator, simu.plans, precision="double")
+        if backend == "CPU":
+            assert np.allclose(
+                E, np.ones((N,), dtype=PRECISION_COMPLEX)
+            ), f"Split step is not unitary. (Backend {backend})"
+        elif backend == "GPU" and NLSE_1d.__CUPY_AVAILABLE__:
+            assert cp.allclose(
+                E, cp.ones((N,), dtype=PRECISION_COMPLEX)
+            ), f"Split step is not unitary. (Backend {backend})"
+
+
 def test_out_field() -> None:
     for backend in ["CPU", "GPU"]:
         simu = NLSE_1d(0, puiss, window, n2, None, L, NX=N, Isat=Isat, backend=backend)
@@ -80,7 +105,6 @@ def test_out_field() -> None:
             E0, simu.delta_z, verbose=False, plot=False, precision="single"
         )
         rho = A.real * A.real + A.imag * A.imag
-        print(rho)
         norm = (rho * simu.delta_X**2).sum(axis=simu._last_axes)
         norm *= c * epsilon_0 / 2
         assert A.shape == (N,), f"Output array has wrong shape. (Backend {backend})"
