### 2024-04-05 42490fa
Tests for NLSE_3d, updated docs
