### 2024-04-05 e8caff9
typo in test_nlse3d
