### 2024-04-05 #6509 @taladjidi
@RMeli all good now ! It was an overflow error due to some edge case in the solver's resolution. Now everything passes 🥳
