### 2024-04-05 4f386e0
Definition of the propagation distance
