+++ NLSE/nlse.py
@@ -476,7 +476,7 @@ class NLSE:
         ax[0].set_title("Intensity")
         ax[0].set_xlabel("x (mm)")
         ax[0].set_ylabel("y (mm)")
-        fig.colorbar(im0, ax=ax[0], shrink=0.6, label="Intensity (W/cm^2)")
+        fig.colorbar(im0, ax=ax[0], shrink=0.6, label=r"Intensity ($W/cm^2$)")
         im1 = ax[1].imshow(
             phi, extent=ext_real, cmap="twilight_shifted", vmin=-np.pi, vmax=np.pi
         )
