### 2024-04-10 #6509 @RMeli
Thanks for the update @taladjidi. Some of the tests you added don't seem very meaningful (`test_build_propagator` just uses the same expression as the implementation), but I'll leave it to the reviewers to comment further. Thanks for adding proper testing.
