### 2024-04-12 #6607 @editorialbot
[REVIEW]: NLSE: A Python package to solve the nonlinear Schrödinger equation
**Submitting author:** @taladjidi (<a href="http://orcid.org/0000-0002-3109-9723">Tangui Aladjidi</a>)
**Repository:** https://github.com/Quantum-Optics-LKB/NLSE
**Branch with paper.md** (empty if default branch):
**Version:** 2.3.0
