+++ README.md
@@ -245,3 +245,9 @@ $$
 i\partial_{t}\psi = -\frac{1}{2}\nabla^2\psi+V\psi+g|\psi|^2\psi.
 $$
 It follows exactly the same conventions as the other classes a part from the fact that since it describes atoms, the units are the "atomic" units (masses in kg, times in s).
+
+
+## Contributing and issues
+
+If you wish to contribute, do not hesitate to create a PR or email me (tangui.aladjidi[at]lkb.upmc.fr).
+If you encounter problems with this software, you can create an issue directly on this repository.
