### 2024-04-16 Quantum-Optics-LKB/NLSE#2 @Abinashbunty
DOI fix for thesis
@taladjidi During my first stage of review for openjournals/joss-reviews#6607, I see that for the reference _Full optical control of quantum fluids of light in hot atomic vapors_, there is no url provided. Maybe you can look into https://theses.hal.science/THESES-SU/tel-04391272v1 and cite the thesis accordingly. This will help in fixing the DOI error that editorialbot is telling about.
