+++ paper/paper.bib
@@ -5,7 +5,8 @@
   abstract = {Quantum fluids of light are a novel and promising platform in the exploration of many-body quantum physics. We study here fluids whose constituants are photons. Their interactions are induced by a hot vapor of Rubidium atoms. In order to describe these fluids, we use a Hamiltonian comprising of three terms: kinetic energy, potential energy and interaction energy. In this thesis, we showcase an all-optical control of each of these terms. We evidenced non-classical correlations induced by the interaction quenches at the interfaces of our non-linear medium. We then explore the control of the atomic response of the vapor, using an easily extendable numerical framework. We then use this framework to study optical pumping, that allows us to realize arbitrary potentials inside the fluid of light. The critical velocity for superfluidity is then mesured, in a scattering experiment against an optical defect using the potential engineering developped in this thesis. Finally, we study vortex-vortex interactions in an experiment showcasing turbulent behavior in a fluid of light as well as a collision experiment between vortices. This paves the way to thermodynamics processes seeded by interactions effects beyondthe mean field.},
   langid = {english},
   school = {Sorbonne Universit{\'e}},
-  address = {Laboratoire Kastler Brossel, Paris, France}
+  address = {Laboratoire Kastler Brossel, Paris, France},
+  url = {https://theses.hal.science/THESES-SU/tel-04391272v1}
 }
 
 @article{glorieuxHotAtomicVapors2023,
