### 2024-04-17 7264586
Update paper.bib
To fix the no doi error
