### 2024-04-17 #6607 @RMeli
@taladjidi I just noticed from the number of DOIs that there are only three references, and they are self-references. Is it correct that there is nothing else to cite from other authors?
