### 2024-04-23 8472836
More bibliography, extended statement of need
