+++ paper/paper.bib
@@ -1,3 +1,60 @@
+@book{pitaevskij_bose-einstein_2016,
+  location  = {Oxford},
+  title     = {Bose-Einstein condensation and superfluidity},
+  isbn      = {978-0-19-875888-4},
+  series    = {International series of monographs on physics},
+  pagetotal = {553},
+  number    = {164},
+  publisher = {Oxford University Press},
+  author    = {Pitaevskij, Lev P. and Stringari, Sandro},
+  year      = {2016}
+}
+
+@article{Schloss2018, 
+  doi = {10.21105/joss.01037}, 
+  url = {https://doi.org/10.21105/joss.01037},
+  year = {2018}, 
+  publisher = {The Open Journal}, 
+  volume = {3},
+  number = {32}, 
+  pages = {1037}, 
+  author = {James R. Schloss and Lee J. O'Riordan}, 
+  title = {GPUE: Graphics Processing Unit Gross--Pitaevskii Equation solver}, 
+  journal = {Journal of Open Source Software} 
+} 
+
+@article{michel_superfluid_2018,
+  title    = {Superfluid motion and drag-force cancellation in a fluid of light},
+  volume   = {9},
+  doi      = {10.1038/s41467-018-04534-9},
+  abstract = {Superfluidity of light enables movement without friction or, in optical terms, diffraction. Here, Michel et al. report the detection of the transition to superfluidity in the flow of a fluid of light past an obstacle in a bulk nonlinear crystal and extract the drag force exerted by the fluid of light.},
+  pages    = {2108},
+  number   = {1},
+  journal  = {Nat. Comm.},
+  author   = {Michel, Claire and Boughdad, Omar and Albert, Mathias and Larré, Pierre-Élie and Bellec, Matthieu},
+  date     = {2018}
+}
+
+@article{carusotto_quantum_2013,
+  title      = {Quantum fluids of light},
+  volume     = {85},
+  issn       = {0034-6861, 1539-0756},
+  url        = {http://arxiv.org/abs/1205.6500},
+  doi        = {10.1103/RevModPhys.85.299},
+  abstract   = {This article reviews recent theoretical and experimental advances in the fundamental understanding and active control of quantum fluids of light in nonlinear optical systems. In presence of effective photon-photon interactions induced by the optical nonlinearity of the medium, a many-photon system can behave collectively as a quantum fluid with a number of novel features stemming from its intrinsically non-equilibrium nature. We present a rich variety of photon hydrodynamical effects that have been recently observed, from the superfluid flow around a defect at low speeds, to the appearance of a Mach-Cherenkov cone in a supersonic flow, to the hydrodynamic formation of topological excitations such as quantized vortices and dark solitons at the surface of large impenetrable obstacles. While our review is mostly focused on a class of semiconductor systems that have been extensively studied in recent years (namely planar semiconductor microcavities in the strong light-matter coupling regime having cavity polaritons as elementary excitations), the very concept of quantum fluids of light applies to a broad spectrum of systems, ranging from bulk nonlinear crystals, to atomic clouds embedded in optical fibers and cavities, to photonic crystal cavities, to superconducting quantum circuits based on Josephson junctions. The conclusive part of our article is devoted to a review of the exciting perspectives to achieve strongly correlated photon gases. In particular, we present different mechanisms to obtain efficient photon blockade, we discuss the novel quantum phases that are expected to appear in arrays of strongly nonlinear cavities, and we point out the rich phenomenology offered by the implementation of artificial gauge fields for photons.},
+  pages      = {299--366},
+  number     = {1},
[…]
