### 2024-04-23 Quantum-Optics-LKB/NLSE#2 @taladjidi
Hi I fixed the bib file in order to add the url. If it's ok with you I mark this issue as closed, thank you for your feedback !
