### 2024-05-03 17826c3
Update nlse.py
