### 2024-05-03 bb21c96
Update nlse_3d.py
