+++ NLSE/nlse_3d.py
@@ -50,7 +50,7 @@ class NLSE_3d(NLSE):
             NY (int, optional): Number of points in the y direction. Defaults to 1024.
             NZ (int, optional): Number of points in the t direction. Defaults to 1024.
             Isat (float): Saturation intensity in W/m^2
-            nl_length (float): Non linear length in m
+            nl_length (float): Non local length in m
             wvl (float): Wavelength in m
             backend (str, optional): "GPU" or "CPU". Defaults to __BACKEND__.
         """
