+++ NLSE/__init__.py
@@ -1,3 +1,17 @@
+"""
+NLSE.
+
+A package for solving the Nonlinear Schrödinger Equation (NLSE) using the
+Split-Step Fourier method.
+"""
+
+__version__ = "2.1.0"
+__author__ = "Tangui Aladjidi"
+__license__ = "MIT"
+__credits__ = "Laboratoire Kastler Brossel, Paris, France"
+__email__ = "tangui.aladjidi@lkb.upmc.fr"
+
+
 from . import utils
 from .nlse import NLSE
 from .nlse_1d import NLSE_1d
