+++ mkdocs-documentation/docs/cnlse.md
@@ -1,0 +1,3 @@
+`CNLSE` is a Python package for solving the complex nonlinear Schrödinger equation (CNLSE) in one and two dimensions. The CNLSE is a partial differential equation that describes the evolution of a complex-valued wave function in a nonlinear medium. The CNLSE is a fundamental equation in the study of nonlinear optics, Bose-Einstein condensates, and other physical systems.
+
+::: NLSE.cnlse
