+++ mkdocs-documentation/docs/reference.md
@@ -1,0 +1,3 @@
+`NLSE` is a Python package for solving the nonlinear Schrödinger equation (NLSE) in one and two dimensions. The NLSE is a partial differential equation that describes the evolution of a complex-valued wave function in a nonlinear medium. The NLSE is a fundamental equation in the study of nonlinear optics, Bose-Einstein condensates, and other physical systems.
+
+::: NLSE.nlse
