### 2024-05-16 1844f4e
test: update test_nlse.py
