+++ mkdocs-documentation/readme.md
@@ -1,0 +1,8 @@
+# Documentation of NLSE
+
+## Installation
+pip install mkdocs
+
+## Dev server
+To launch the dev server : mkdocs serve
+Then go to [Here](http://127.0.0.1:8000/)
