+++ mkdocs-documentation/mkdocs.yml
@@ -11,6 +11,7 @@ plugins:
 
 nav:
   - Home: index.md
+  - Install: install.md
   - NLSE: reference.md
   - Coupled NLSE: cnlse.md
   - Kernels: kernels.md
