+++ NLSE/__init__.py
@@ -19,3 +19,4 @@ from .nlse_3d import NLSE_3d
 from .cnlse import CNLSE
 from .cnlse_1d import CNLSE_1d
 from .gpe import GPE
+from .ddgpe import DDGPE
