+++ NLSE/ddgpe.py
@@ -1,8 +1,8 @@
 from .cnlse import CNLSE
 import numpy as np
 import matplotlib.pyplot as plt
-from scipy.constants import hbar
 import pyfftw
+from typing import Union
 from .utils import __BACKEND__, __CUPY_AVAILABLE__
 
 if __CUPY_AVAILABLE__:
@@ -14,19 +14,20 @@ class DDGPE(CNLSE):
 
     def __init__(
         self,
-        alpha: float,
+        gamma: float,
         puiss: float,
         window: float,
-        n2: float,
+        g: float,
         omega: float,
-        V: np.ndarray,
-        L: float,
-        n12: float = 0,
+        T: float,
+        m: float = np.inf,
+        m2: float = 2 * np.pi / 800e-9,
+        V: np.ndarray = None,
+        g12: float = 0,
         NX: int = 1024,
         NY: int = 1024,
         Isat: float = np.inf,
         nl_length: float = 0,
-        wvl: float = 780e-9,
         backend: str = __BACKEND__,
     ) -> object:
         """Instantiates the class with all the relevant physical parameters
@@ -53,29 +54,34 @@ class DDGPE(CNLSE):
         """
 
         super().__init__(
-            alpha=alpha,
+            alpha=gamma,
             puiss=puiss,
             window=window,
-            n2=n2,
-            n12=n12,
+            n2=g,
+            n12=g12,
             V=V,
-            L=L,
+            L=T,
             NX=NX,
             NY=NY,
             Isat=Isat,
             nl_length=nl_length,
-            wvl=wvl,
+            wvl=2 * np.pi / m + 1e-30,
             omega=omega,
             backend=backend,
         )
-        self.alpha *= hbar
-        self.omega *= hbar
-        self.n2 *= hbar
-        self.n22 = 0
-        # k = 2 * np.pi / self.wvl = m
+        self.g = self.n2
+        self.g12 = self.n12
+        self.g2 = self.n22
+        self.m = m
+        self.m2 = m2
+        self.omega0_photon = 0
+        self.gamma = gamma
+        self.gamma2 = self.gamma
 
     @staticmethod
-    def add_noise(simu: object, A: np.ndarray, t: float, i: int, *args, **kwargs):
+    def add_noise(
+        simu: object, A: np.ndarray, t: float, i: int, *args, **kwargs
+    ) -> None:
         """Add noise to the propagation step.
 
         Follows the callback convention of NLSE.
@@ -90,12 +96,340 @@ class DDGPE(CNLSE):
         # will certainly benefit from kernel fusion
         pass
 
-    def _build_propagator(self) -> np.ndarray:
-        # kinetic terms for the exciton dispersion and polariton dispersion
+    @staticmethod
+    def pump(simu: object, A: np.ndarray, t: float, i: int, *args, **kwargs) -> None:
+        """Add the pump laser.
 
-        # propagator1 = super()._build_propagator()
-        # propagator2 = np.exp(
-        #     -1j * 0.5 * (self.Kxx**2 + self.Kyy**2) / self.k2 * self.delta_z
-        # ).astype(np.complex64)
-        # return np.array([propagator1, propagator2])
+        Args:
+            simu (object): _description_
+            A (np.ndarray): _description_
+            t (float): _description_
+            i (int): _description_
+        """
         pass
+
+    def _build_propagator(self) -> np.ndarray:
+        """Build the propagators.
+
+        Args:
+            omega (float): Arbitrary frequency offset.
+        Returns:
+            np.ndarray: A tuple of linear propagators for each component.
+        """
+        propagator1 = np.exp(
+            -1j * 0.5 * (self.Kxx**2 + self.Kyy**2) / self.m * self.delta_z
+        ).astype(np.complex64)
+        propagator2 = np.exp(
+            -1j
[…]
