### 2024-05-23 5744d90
some work done on ddgpe and test_ddgpe
