+++ NLSE/ddgpe.py
@@ -20,8 +20,9 @@ class DDGPE(CNLSE):
         g: float,
         omega: float,
         T: float,
-        m: float = np.inf,
-        m2: float = 2 * np.pi / 800e-9,
+        omega_exc: float,
+        omega_cav: float,
+        k_z: float,
         V: np.ndarray = None,
         g12: float = 0,
         NX: int = 1024,
@@ -65,22 +66,22 @@ class DDGPE(CNLSE):
             NY=NY,
             Isat=Isat,
             nl_length=nl_length,
-            wvl=2 * np.pi / m + 1e-30,
+            wvl=1e-30,
             omega=omega,
             backend=backend,
         )
         self.g = self.n2
         self.g12 = self.n12
         self.g2 = self.n22
-        self.m = m
-        self.m2 = m2
-        self.omega0_photon = 0
+        self.omega_exc = omega_exc
+        self.omega_cav = omega_cav
+        self.k_z = k_z
         self.gamma = gamma
         self.gamma2 = self.gamma
-
+        
     @staticmethod
     def add_noise(
-        simu: object, A: np.ndarray, t: float, i: int, *args, **kwargs
+        simu: object, A: np.ndarray, t: float, i: int, noise:float = 0, noise1:float = 0, **kwargs
     ) -> None:
         """Add noise to the propagation step.
 
@@ -92,13 +93,32 @@ class DDGPE(CNLSE):
             t (float): Propagation time in s.
             i (int): Propagation step.
         """
-        # do something like A += np.random.normal(0, 1e-3, A.shape)
-        # will certainly benefit from kernel fusion
-        pass
+        if simu.backend == "GPU" and simu.__CUPY_AVAILABLE__:
+            rand1 = cp.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX), dtype = np.float64) + 1j * cp.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX), dtype = np.float64)
+            rand2 = cp.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX), dtype = np.float64) + 1j * cp.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX), dtype = np.float64)
+            A[...,0,:,:] += noise * cp.sqrt(simu.gamma / (4 * (simu.delta_X*simu.delta_Y))) * rand1
+            A[...,1,:,:] += noise * cp.sqrt((simu.gamma2) / (4 * (simu.delta_X*simu.delta_Y))) * rand2
 
+        else:
+            rand1 = np.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX)) + 1j * np.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX))
+            rand2 = np.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX)) + 1j * np.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX))
+            A[...,0,:,:] += noise * np.sqrt(simu.gamma / (4 * (simu.delta_X*simu.delta_Y))) * rand1
+            A[...,1,:,:] += noise * np.sqrt((simu.gamma2) / (4 * (simu.delta_X*simu.delta_Y))) * rand2
+    
     @staticmethod
-    def pump(simu: object, A: np.ndarray, t: float, i: int, *args, **kwargs) -> None:
-        """Add the pump laser.
+    def laser_excitation(simu: object, 
+            A: np.ndarray, 
+            t: float, 
+            i: int, 
+            F_pump: float,
+            F_pump_r: cp.ndarray, 
+            F_pump_t: float, 
+            F_probe: float = 0, 
[…]
