+++ NLSE/kernels_cpu.py
@@ -73,7 +73,8 @@ def nl_prop_c(
     V: np.ndarray,
     g11: float,
     g12: float,
-    Isat: float,
+    Isat1: float,
+    Isat2: float,
 ) -> None:
     """A fused kernel to apply real space terms
     Args:
@@ -85,18 +86,20 @@ def nl_prop_c(
         V (cp.ndarray): Potential
         g11 (float): Intra-component interactions
         g12 (float): Inter-component interactions
-        Isat (float): Saturation parameter of intra-component interaction
+        Isat1 (float): Saturation parameter of first component
+        Isat2 (float): Saturation parameter of second component
     """
     A1 = A1.ravel()
     A_sq_1 = A_sq_1.ravel()
     A_sq_2 = A_sq_2.ravel()
     for i in numba.prange(A1.size):
         # Saturation parameter
-        sat = 1 / (1 + A_sq_1[i] / Isat)
+        sat1 = 1 / (1 + A_sq_1[i] / Isat1)
+        sat2 = 1 / (1 + A_sq_2[i] / Isat2)
         # Losses
-        arg = -alpha / 2 * sat
+        arg = -alpha / 2 * sat1
         # Interactions
-        arg += 1j * (g11 * A_sq_1[i] * sat + g12 * A_sq_2[i])
+        arg += 1j * (g11 * A_sq_1[i] * sat1 + g12 * A_sq_2[i] * sat2)
         # Potential
         arg += 1j * V[i]
         A1[i] *= np.exp(dz * arg)
@@ -111,7 +114,8 @@ def nl_prop_without_V_c(
     alpha: float,
     g11: float,
     g12: float,
-    Isat: float,
+    Isat1: float,
+    Isat2: float,
 ) -> None:
     """A fused kernel to apply real space terms
     Args:
@@ -122,18 +126,20 @@ def nl_prop_without_V_c(
         alpha (float): Losses
         g11 (float): Intra-component interactions
         g12 (float): Inter-component interactions
-        Isat (float): Saturation parameter of intra-component interaction
+        Isat1 (float): Saturation parameter of first component
+        Isat2 (float): Saturation parameter of second component
     """
     A1 = A1.ravel()
     A_sq_1 = A_sq_1.ravel()
     A_sq_2 = A_sq_2.ravel()
     for i in numba.prange(A1.size):
         # Saturation parameter
-        sat = 1 / (1 + A_sq_1[i] / Isat)
+        sat1 = 1 / (1 + A_sq_1[i] / Isat1)
+        sat2 = 1 / (1 + A_sq_2[i] / Isat2)
         # Losses
-        arg = -alpha / 2 * sat
+        arg = -alpha / 2 * sat1
         # Interactions
-        arg += 1j * (g11 * A_sq_1[i] * sat + g12 * A_sq_2[i])
+        arg += 1j * (g11 * A_sq_1[i] * sat1 + g12 * A_sq_2[i] * sat2)
         A1[i] *= np.exp(dz * arg)
 
 
