+++ tests/test_ddgpe.py
@@ -1,5 +1,6 @@
 from NLSE import DDGPE
 import numpy as np
+import matplotlib.pyplot as plt
 
 if DDGPE.__CUPY_AVAILABLE__:
     import cupy as cp
@@ -7,7 +8,7 @@ if DDGPE.__CUPY_AVAILABLE__:
 def turn_on(
     F_laser_t: np.ndarray, 
     time: np.ndarray, 
-    t_up=200
+    t_up=10,
 ):
     """A function to turn on the pump more or less adiabatically
 
@@ -16,48 +17,63 @@ def turn_on(
         time (np.ndarray):  array with the value of the time at each discretized step
         t_up (int, optional): time taken to reach the maximum intensity (=F). Defaults to 200.
     """
-    F_laser_t[time < t_up] = np.exp(-1 * (time[time < t_up] - t_up)**2 / (t_up / 2))
+    F_laser_t[time < t_up] = np.exp(-1 * (time[time < t_up] - t_up)**2 / (t_up / 2)**2)
     F_laser_t[time >= t_up] = 1
 
-
-
+def callback_sample(simu: DDGPE, A: np.ndarray, z: float, i: int, save_every: int, sample1: list, sample2: list, sample3: list) -> None:
+    if i % save_every == 0:
+        sum_exc = np.sum(np.abs(A[..., 0, :, :].get())**2)
+        sum_cav = np.sum(np.abs(A[..., 1, :, :].get())**2)
+        sum_tot = np.sum(np.abs(A[..., :, :, :].get())**2)
+        sample1.append(sum_exc)
+        sample2.append(sum_cav)
+        sample3.append(sum_tot)
+        
 def main():
     
     h_bar = 0.654 # (meV*ps)
     c = 2.9979*1e2 # (um/ps)
-    omega = 0.5*5.07/h_bar # (meV/h_bar) linear coupling (Rabi split)
-    g0 = (1e-2) /h_bar  # (frequency/density) (meV/hbar)/(1/um^2) nonlinear coupling constant 
-    omega_exc = 1484.44 /h_bar # (meV/h_bar) exciton energy measured from the cavity energy #-0.5
-    omega_cav = 1482.76 /h_bar # (meV/h_bar) cavity energy at k=0  original value: 1482.76 /h_bar
-    omega_lp = (omega_exc + omega_cav) / 2 - 0.5 * np.sqrt((omega_exc - omega_cav) ** 2 + 4 * (omega) ** 2)
-    detuning = 0.17/h_bar
-    omega_pump = omega_lp + detuning
-    omega_exc-=omega_pump
-    omega_cav-=omega_pump
+    omega = 5.07 / h_bar # (meV/h_bar) linear coupling (Rabi split)
+    omega_exc = 1484.44 / h_bar # (meV/h_bar) exciton energy 
+    omega_cav = 1482.76 / h_bar # (meV/h_bar) cavity energy 
+    omega_lp = (omega_exc + omega_cav) / 2 - 0.5 * np.sqrt((omega_exc - omega_cav) ** 2 + (omega) ** 2)
+    detuning = 0.17 / h_bar
     k_z = 27
-    gamma = 10 * 0.07 / h_bar
+    gamma = 0*0.07 / h_bar
     puiss = 1
     waist = 50
-    window = 4*waist
-    g = 1e-2
-    T = 6e2
+    window = 256
+    g = 1e-2 / h_bar
+    T = 100
     
-    dd = DDGPE(gamma, puiss, window, g, omega, T, omega_exc, omega_cav, k_z, NX=256, NY=256)
[…]
