### 2024-05-27 54c9a1c
Correct Rabi term
