+++ NLSE/ddgpe.py
@@ -80,12 +80,20 @@ class DDGPE(CNLSE):
         self.omega_exc = omega_exc
         self.omega_cav = omega_cav
         self.detuning = detuning
-        omega_lp = (omega_exc + omega_cav) / 2 - 0.5 * np.sqrt((omega_exc - omega_cav) ** 2 + (omega) ** 2)
+        omega_lp = (omega_exc + omega_cav) / 2 - 0.5 * np.sqrt(
+            (omega_exc - omega_cav) ** 2 + (omega) ** 2
+        )
         self.omega_pump = omega_lp + detuning
 
     @staticmethod
     def add_noise(
-        simu: object, A: np.ndarray, t: float, i: int, noise:float = 0, noise1:float = 0, **kwargs
+        simu: object,
+        A: np.ndarray,
+        t: float,
+        i: int,
+        noise: float = 0,
+        noise1: float = 0,
+        **kwargs,
     ) -> None:
         """Add noise to the propagation step.
 
@@ -98,30 +106,63 @@ class DDGPE(CNLSE):
             i (int): Propagation step.
         """
         if simu.backend == "GPU" and simu.__CUPY_AVAILABLE__:
-            rand1 = cp.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX), dtype = np.float64) + 1j * cp.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX), dtype = np.float64)
-            rand2 = cp.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX), dtype = np.float64) + 1j * cp.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX), dtype = np.float64)
-            A[...,0,:,:] += noise * cp.sqrt(simu.gamma / (4 * (simu.delta_X*simu.delta_Y))) * rand1
-            A[...,1,:,:] += noise * cp.sqrt((simu.gamma2) / (4 * (simu.delta_X*simu.delta_Y))) * rand2
+            rand1 = cp.random.normal(
+                loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX), dtype=np.float64
+            ) + 1j * cp.random.normal(
+                loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX), dtype=np.float64
+            )
+            rand2 = cp.random.normal(
+                loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX), dtype=np.float64
+            ) + 1j * cp.random.normal(
+                loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX), dtype=np.float64
+            )
+            A[..., 0, :, :] += (
+                noise
+                * cp.sqrt(simu.gamma / (4 * (simu.delta_X * simu.delta_Y)))
+                * rand1
+            )
+            A[..., 1, :, :] += (
+                noise
+                * cp.sqrt((simu.gamma2) / (4 * (simu.delta_X * simu.delta_Y)))
+                * rand2
+            )
 
         else:
-            rand1 = np.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX)) + 1j * np.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX))
-            rand2 = np.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX)) + 1j * np.random.normal(loc = 0, scale = simu.delta_z, size = (simu.NY, simu.NX))
-            A[...,0,:,:] += noise * np.sqrt(simu.gamma / (4 * (simu.delta_X*simu.delta_Y))) * rand1
[…]
