### 2024-05-27 dc4dcd6
Good Rabi term
