+++ NLSE/ddgpe.py
@@ -451,9 +451,7 @@ class DDGPE(CNLSE):
                     self.I_sat,
                 )
             if self.omega is not None:
-                A1_old = A1.copy()
-                self._kernels.rabi_coupling(A1, A2, self.delta_z, self.omega / 2)
-                self._kernels.rabi_coupling(A2, A1_old, self.delta_z, self.omega / 2)
+                self._kernels.rabi_coupling(A, self.delta_z, self.omega / 2)
 
     def plot_field(self, A_plot: np.ndarray, t: float) -> None:
         """Plot the field for monitoring.
