### 2024-05-27 8c1d71f
kernels_cpu with new Rabi
