### 2024-05-27 #6607 @taladjidi
Hi @obliviateandsurrender, thanks for the review and your thoughtful remarks. I recently did a tutorial on this piece of software to my team so I have a jupyter notebook on hand, it should not take to much back and forth ! @RMeli  I'll try to address everything by the end of the week.
Thank you both again for your time
