+++ mkdocs-documentation/docs/index.md
@@ -1,0 +1,6 @@
+# Welcome to NLSE
+
+`NLSE` is a Python package for solving the nonlinear Schrödinger equation (NLSE) in one and two dimensions. The NLSE is a partial differential equation that describes the evolution of a complex-valued wave function in a nonlinear medium. The NLSE is a fundamental equation in the study of nonlinear optics, Bose-Einstein condensates, and other physical systems.
+
+This package provides utility to easily simulate all sorts of non linear Schrödinger equations. It uses a [split-step spectral scheme](https://en.wikipedia.org/wiki/Split-step_method) to solve the equations.
+
