+++ mkdocs-documentation/docs/reference.md
@@ -1,0 +1,6 @@
+In this page you will find the reference documentation for the NLSE package.
+We describe the classes and functions that are available in the package.
+The class NLSE is the main class available in the package.
+
+
+::: NLSE.nlse.NLSE
