+++ mkdocs-documentation/docs/kernels.md
@@ -1,0 +1,17 @@
+## Kernels
+
+### CPU
+The CPU kernel in the `NLSE` package is responsible for solving the nonlinear Schrödinger equation using CPU resources. It utilizes the computational power of the central processing unit (CPU) to perform the necessary calculations.
+
+To use the CPU kernel, you can call the appropriate functions or methods provided by the `NLSE` package. These functions will internally invoke the CPU-specific implementation to solve the NLSE.
+
+The CPU kernel is suitable for small to medium-sized problems or when GPU resources are not available. It provides a reliable and efficient solution for solving the NLSE on CPU-based systems.
+
+If you have specific code or implementation details related to the CPU kernel that you would like assistance with, please provide more information and I'll be happy to help.
+
+::: NLSE.kernels_cpu
+
+### GPU (CUDA)
+The GPU kernel in the `NLSE` package is responsible for solving the nonlinear Schrödinger equation using GPU resources. It utilizes the computational power of the graphics processing unit (GPU) to perform the necessary calculations.
+
+::: NLSE.kernels_gpu
