+++ mkdocs-documentation/docs/reference.md
@@ -1,12 +1,6 @@
-<<<<<<< HEAD
 In this page you will find the reference documentation for the NLSE package.
 We describe the classes and functions that are available in the package.
 The class NLSE is the main class available in the package.
 
 
 ::: NLSE.nlse.NLSE
-=======
-`NLSE` is a Python package for solving the nonlinear Schrödinger equation (NLSE) in one and two dimensions. The NLSE is a partial differential equation that describes the evolution of a complex-valued wave function in a nonlinear medium. The NLSE is a fundamental equation in the study of nonlinear optics, Bose-Einstein condensates, and other physical systems.
-
-::: NLSE.nlse
->>>>>>> b2d5ce201112718ea8f7ad57179765703d28a237
