+++ mkdocs-documentation/mkdocs.yml
@@ -11,7 +11,6 @@ plugins:
 
 nav:
   - Home: index.md
-<<<<<<< HEAD
   - Getting started:
     - Installation: 'installation.md'
     - Basic Usage: 'basic_usage.md'
@@ -27,10 +26,4 @@ nav:
     - NLSE: reference.md
     - Coupled NLSE: cnlse.md
     - Kernels: kernels.md
-=======
-  - Install: install.md
-  - NLSE: reference.md
-  - Coupled NLSE: cnlse.md
-  - Kernels: kernels.md
->>>>>>> b2d5ce201112718ea8f7ad57179765703d28a237
 
