### 2024-05-28 bbaa45b
Latex support
