+++ mkdocs-documentation/mkdocs.yml
@@ -27,3 +27,11 @@ nav:
     - Coupled NLSE: cnlse.md
     - Kernels: kernels.md
 
+markdown_extensions:
+  - pymdownx.arithmatex:
+      generic: true
+
+extra_javascript:
+  - javascripts/mathjax.js
+  - https://polyfill.io/v3/polyfill.min.js?features=es6
+  - https://unpkg.com/mathjax@3/es5/tex-mml-chtml.js
