### 2024-05-28 f3f8447
Big clean up, proper saturation
