### 2024-05-28 3fd149d
a_sq as argument of split_step for better control over memory
