+++ NLSE/kernels_gpu.py
@@ -75,8 +75,7 @@ def nl_prop_c(
     V: cp.ndarray,
     g11: float,
     g12: float,
-    Isat1: float,
-    Isat2: float,
+    Isat: float,
 ) -> None:
     """A fused kernel to apply real space terms
     Args:
@@ -92,12 +91,11 @@ def nl_prop_c(
         Isat2 (float): Saturation parameter of second component
     """
     # Saturation parameter
-    sat1 = 1 / (1 + A_sq_1 / Isat1)
-    sat2 = 1 / (1 + A_sq_2 / Isat2)
+    sat = 1 / (1 + (A_sq_1 + A_sq_2) / Isat)
     # Interactions
-    arg = 1j * (g11 * A_sq_1 * sat1 + g12 * A_sq_2 * sat2)
+    arg = 1j * (g11 * A_sq_1 * sat + g12 * A_sq_2 * sat)
     # Losses
-    arg += -alpha / 2 * sat1
+    arg += -alpha / 2 * sat
     # Potential
     arg += 1j * V
     arg *= dz
@@ -114,8 +112,7 @@ def nl_prop_without_V_c(
     alpha: float,
     g11: float,
     g12: float,
-    Isat1: float,
-    Isat2: float,
+    Isat: float,
 ) -> None:
     """A fused kernel to apply real space terms
     Args:
@@ -130,12 +127,11 @@ def nl_prop_without_V_c(
         Isat2 (float): Saturation parameter of second component
     """
     # Saturation parameter
-    sat1 = 1 / (1 + A_sq_1 / Isat1)
-    sat2 = 1 / (1 + A_sq_2 / Isat2)
+    sat = 1 / (1 + A_sq_1 / Isat)
     # Interactions
-    arg = 1j * (g11 * A_sq_1 * sat1 + g12 * A_sq_2 * sat2)
+    arg = 1j * (g11 * A_sq_1 * sat + g12 * A_sq_2 * sat)
     # Losses
-    arg += -alpha / 2 * sat1
+    arg += -alpha / 2 * sat
     arg *= dz
     cp.exp(arg, out=arg)
     A1 *= arg
@@ -191,4 +187,4 @@ def square_mod(A: cp.ndarray, A_sq: cp.ndarray) -> None:
     Returns:
         None
     """
-    A_sq[:] = A.real * A.real + A.imag * A.imag
+    A_sq[:] = cp.abs(A) ** 2
