### 2024-05-28 fa7d1b7
Documentation, clean up of DDGPE
