+++ NLSE/ddgpe.py
@@ -84,6 +84,10 @@ class DDGPE(CNLSE):
             (omega_exc - omega_cav) ** 2 + (omega) ** 2
         )
         self.omega_pump = omega_lp + detuning
+        if self.backend == "GPU" and self.__CUPY_AVAILABLE__:
+            self._random = cp.random.normal
+        else:
+            self._random = np.random.normal
 
     @staticmethod
     def add_noise(
@@ -92,8 +96,6 @@ class DDGPE(CNLSE):
         t: float,
         i: int,
         noise: float = 0,
-        noise1: float = 0,
-        **kwargs,
     ) -> None:
         """Add noise to the propagation step.
 
@@ -105,49 +107,18 @@ class DDGPE(CNLSE):
             t (float): Propagation time in s.
             i (int): Propagation step.
         """
-        if simu.backend == "GPU" and simu.__CUPY_AVAILABLE__:
-            rand1 = cp.random.normal(
-                loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX), dtype=np.float64
-            ) + 1j * cp.random.normal(
-                loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX), dtype=np.float64
-            )
-            rand2 = cp.random.normal(
-                loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX), dtype=np.float64
-            ) + 1j * cp.random.normal(
-                loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX), dtype=np.float64
-            )
-            A[..., 0, :, :] += (
-                noise
-                * cp.sqrt(simu.gamma / (4 * (simu.delta_X * simu.delta_Y)))
-                * rand1
-            )
-            A[..., 1, :, :] += (
-                noise
-                * cp.sqrt((simu.gamma2) / (4 * (simu.delta_X * simu.delta_Y)))
-                * rand2
-            )
-
-        else:
-            rand1 = np.random.normal(
-                loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX)
-            ) + 1j * np.random.normal(
-                loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX)
-            )
-            rand2 = np.random.normal(
-                loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX)
-            ) + 1j * np.random.normal(
-                loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX)
-            )
-            A[..., 0, :, :] += (
-                noise
-                * np.sqrt(simu.gamma / (4 * (simu.delta_X * simu.delta_Y)))
-                * rand1
-            )
-            A[..., 1, :, :] += (
-                noise
-                * np.sqrt((simu.gamma2) / (4 * (simu.delta_X * simu.delta_Y)))
-                * rand2
-            )
+        rand1 = simu.normal(
+            loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX)
+        ) + 1j * simu.normal(loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX))
+        rand2 = simu.normal(
+            loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX)
+        ) + 1j * simu.normal(loc=0, scale=simu.delta_z, size=(simu.NY, simu.NX))
+        A[..., 0, :, :] += (
+            noise * np.sqrt(simu.gamma / (4 * (simu.delta_X * simu.delta_Y))) * rand1
+        )
+        A[..., 1, :, :] += (
+            noise * np.sqrt((simu.gamma2) / (4 * (simu.delta_X * simu.delta_Y))) * rand2
+        )
 
[…]
