### 2024-05-28 5dba846
Update kernels_gpu.py
Typo in saturation param
