### 2024-05-28 594b2f1
Update kernels_gpu.py
Typo in saturation param
