+++ NLSE/kernels_gpu.py
@@ -75,7 +75,8 @@ def nl_prop_c(
     V: cp.ndarray,
     g11: float,
     g12: float,
-    Isat: float,
+    Isat1: float,
+    Isat2: float,
 ) -> None:
     """A fused kernel to apply real space terms
     Args:
@@ -91,7 +92,8 @@ def nl_prop_c(
         Isat2 (float): Saturation parameter of second component
     """
     # Saturation parameter
-    sat = 1 / (1 + (A_sq_1 + A_sq_2) / Isat)
+    # sat = 1 / (1 + (A_sq_1 + A_sq_2) / Isat)
+    sat = 1 / (1 + A_sq_1 * 1 / Isat1 + A_sq_2 * 1 / Isat2)
     # Interactions
     arg = 1j * (g11 * A_sq_1 * sat + g12 * A_sq_2 * sat)
     # Losses
@@ -112,7 +114,8 @@ def nl_prop_without_V_c(
     alpha: float,
     g11: float,
     g12: float,
-    Isat: float,
+    Isat1: float,
+    Isat2: float,
 ) -> None:
     """A fused kernel to apply real space terms
     Args:
@@ -127,7 +130,8 @@ def nl_prop_without_V_c(
         Isat2 (float): Saturation parameter of second component
     """
     # Saturation parameter
-    sat = 1 / (1 + (A_sq_1 + A_sq_2) / Isat)
+    # sat = 1 / (1 + (A_sq_1 + A_sq_2) / Isat)
+    sat = 1 / (1 + A_sq_1 * 1 / Isat1 + A_sq_2 * 1 / Isat2)
     # Interactions
     arg = 1j * (g11 * A_sq_1 * sat + g12 * A_sq_2 * sat)
     # Losses
