+++ mkdocs-documentation/mkdocs.yml
@@ -8,24 +8,25 @@ plugins:
       handlers:
         python:
           paths: [..]
+  - mkdocs-jupyter
 
 nav:
   - Home: index.md
   - Getting started:
-    - Installation: 'installation.md'
-    - Basic Usage: 'basic_usage.md'
-    
-    - Structure and Inheritance: 'structure_and_inheritance.md'
-    - Contributing: 'contributing.md'
+      - Installation: "installation.md"
+      - Basic Usage: "basic_usage.md"
+
+      - Structure and Inheritance: "structure_and_inheritance.md"
+      - Contributing: "contributing.md"
   - Tutorial:
-    - What is the Physics problem ?: 'physics_problem.md'
-    - Tutorial on : 'tutorial.md'
+      - What is the Physics problem ?: "physics_problem.md"
+      - Tutorial on: "tutorial.md"
   - How to:
-    - How to ?: 'how_to.md'
+      - How to ?: "how_to.md"
   - Reference:
-    - NLSE: reference.md
-    - Coupled NLSE: cnlse.md
-    - Kernels: kernels.md
+      - NLSE: reference.md
+      - Coupled NLSE: cnlse.md
+      - Kernels: kernels.md
 
 markdown_extensions:
   - pymdownx.arithmatex:
