+++ NLSE/nlse.py
@@ -38,6 +38,7 @@ class NLSE:
     """A class to solve NLSE"""
 
     __CUPY_AVAILABLE__ = __CUPY_AVAILABLE__
+    __PYOPENCL_AVAILABLE__ = __PYOPENCL_AVAILABLE__
 
     def __init__(
         self,
@@ -224,7 +225,7 @@ class NLSE:
             A = cp.empty_like(E_in)
             E_in = cp.asarray(E_in)
         elif self.backend == "CL" and self.__PYOPENCL_AVAILABLE__:
-            A = cla.empty_like(E_in)
+            A = cla.empty(self._cl_queue, E_in.shape, E_in.dtype)
             E_in = cla.to_device(self._cl_queue, E_in)
         else:
             A = pyfftw.empty_aligned(
@@ -232,11 +233,20 @@ class NLSE:
             )
         if normalize:
             # normalization of the field
-            integral = (
-                (E_in.real * E_in.real + E_in.imag * E_in.imag)
-                * self.delta_X
-                * self.delta_Y
-            ).sum(axis=self._last_axes)
+            if self.backend == "CL" and self.__PYOPENCL_AVAILABLE__:
+                arr = E_in.real * E_in.real + E_in.imag * E_in.imag
+                arr *= self.delta_X * self.delta_Y
+                integral = cla.sum(
+                    arr,
+                    dtype=arr.dtype,
+                    queue=self._cl_queue,
+                )
+            else:
+                integral = (
+                    (E_in.real * E_in.real + E_in.imag * E_in.imag)
+                    * self.delta_X
+                    * self.delta_Y
+                ).sum(axis=self._last_axes)
             integral *= c * epsilon_0 / 2
             E_00 = (self.puiss / integral) ** 0.5
             A[:] = (E_00.T * E_in.T).T
@@ -246,18 +256,32 @@ class NLSE:
         """
         Send arrays to GPU.
         """
-        if self.V is not None:
-            self.V = cp.asarray(self.V)
-        self.nl_profile = cp.asarray(self.nl_profile)
-        self.propagator = cp.asarray(self.propagator)
-        # for broadcasting of parameters in case they are
-        # not already on the GPU
-        if isinstance(self.n2, np.ndarray):
-            self.n2 = cp.asarray(self.n2)
-        if isinstance(self.alpha, np.ndarray):
-            self.alpha = cp.asarray(self.alpha)
-        if isinstance(self.I_sat, np.ndarray):
-            self.I_sat = cp.asarray(self.I_sat)
+        if self.backend == "GPU" and self.__CUPY_AVAILABLE__:
+            if self.V is not None:
+                self.V = cp.asarray(self.V)
+            self.nl_profile = cp.asarray(self.nl_profile)
+            self.propagator = cp.asarray(self.propagator)
+            # for broadcasting of parameters in case they are
+            # not already on the GPU
+            if isinstance(self.n2, np.ndarray):
+                self.n2 = cp.asarray(self.n2)
+            if isinstance(self.alpha, np.ndarray):
+                self.alpha = cp.asarray(self.alpha)
+            if isinstance(self.I_sat, np.ndarray):
+                self.I_sat = cp.asarray(self.I_sat)
+        elif self.backend == "CL" and self.__PYOPENCL_AVAILABLE__:
+            if self.V is not None:
+                self.V = cla.to_device(self._cl_queue, self.V)
+            self.nl_profile = cla.to_device(self._cl_queue, self.nl_profile)
+            self.propagator = cla.to_device(self._cl_queue, self.propagator)
+            # for broadcasting of parameters in case they are
+            # not already on the GPU
[…]
