+++ mkdocs-documentation/readme.md
@@ -37,9 +37,10 @@ mkdocs, version 1.2.0 from /usr/local/lib/python3.8/site-packages/mkdocs (Python
 
 ### Extra packages
 
-You need to install packages to support the markdown extensions : [mkdocstrings](https://pypi.org/project/mkdocstrings/) and [mkdocs-jupyter](https://github.com/danielfrg/mkdocs-jupyter) used in the documentation. You can install these using pip:
+You need to install packages to support the markdown extensions : [mkdocsmateral](https://github.com/squidfunk/mkdocs-material), [mkdocstrings](https://pypi.org/project/mkdocstrings/) and [mkdocs-jupyter](https://github.com/danielfrg/mkdocs-jupyter) used in the documentation. You can install these using pip:
 
 ```
+pip install mkdocs-material
 pip install mkdocs-jupyter
 pip install mkdocstrings
 ```
