### 2024-05-28 36ffcc6
docs: update mkdocs-documentation/readme.md
