### 2024-05-28 61ef51c
update how_to.md, reference.md and test_gpe.py
