+++ mkdocs-documentation/docs/how_to.md
@@ -1,0 +1,8 @@
+# Benchmarking
+
+If you want to benchmark the performance of the code you could use the function `tests/benchmarks.py` which is a simple script that runs the code multiple times and calculates the average time taken to run the code.
+
+::: tests.test_gpe
+    options:
+        show_source: true
+
