### 2024-05-29 e180174
big cleanup and more testing
