+++ tests/test_cnlse_1d.py
@@ -49,7 +49,20 @@ def test_prepare_output_array() -> None:
             A = np.ones((2, N), dtype=PRECISION_COMPLEX)
         elif backend == "GPU" and CNLSE_1d.__CUPY_AVAILABLE__:
             A = cp.ones((2, N), dtype=PRECISION_COMPLEX)
-        out = simu._prepare_output_array(A, normalize=True)
+        out, out_sq = simu._prepare_output_array(A, normalize=True)
+        assert (
+            out.flags.c_contiguous
+        ), f"Output array is not C-contiguous. (Backend {backend})"
+        assert (
+            out_sq.flags.c_contiguous
+        ), f"Output array is not C-contiguous. (Backend {backend})"
+        if backend == "CPU":
+            assert (
+                out.flags.aligned
+            ), f"Output array is not aligned. (Backend {backend})"
+            assert (
+                out_sq.flags.aligned
+            ), f"Output array is not aligned. (Backend {backend})"
         integral = ((out.real * out.real + out.imag * out.imag) * simu.delta_X**2).sum(
             axis=simu._last_axes
         )
@@ -88,8 +101,7 @@ def test_split_step() -> None:
         simu.delta_z = 0
         simu.propagator = simu._build_propagator()
         E = np.ones((2, N), dtype=PRECISION_COMPLEX)
-        A = simu._prepare_output_array(E, normalize=False)
-        A_sq = A.copy().real
+        A, A_sq = simu._prepare_output_array(E, normalize=False)
         simu.plans = simu._build_fft_plan(A)
         simu.propagator = simu._build_propagator()
         if backend == "GPU" and CNLSE_1d.__CUPY_AVAILABLE__:
