### 2024-05-29 2c5d33a
more testing, plotting
