+++ NLSE/nlse_3d.py
@@ -111,14 +111,14 @@ class NLSE_3d(NLSE):
             A_sq (np.ndarray): Output field modulus squared array
         """
         if self.backend == "GPU" and self.__CUPY_AVAILABLE__:
-            A = cp.empty_like(E_in)
-            A_sq = cp.empty_like(A, dtype=A.real.dtype)
+            A = cp.zeros_like(E_in)
+            A_sq = cp.zeros_like(A, dtype=A.real.dtype)
             E_in = cp.asarray(E_in)
         else:
-            A = pyfftw.empty_aligned(
+            A = pyfftw.zeros_aligned(
                 E_in.shape, dtype=E_in.dtype, n=pyfftw.simd_alignment
             )
-            A_sq = np.empty_like(A, dtype=A.real.dtype)
+            A_sq = np.zeros_like(A, dtype=A.real.dtype)
         if normalize:
             # normalization of the field
             integral = (
@@ -145,7 +145,7 @@ class NLSE_3d(NLSE):
                 A_plot = A_plot[0]
         if self.__CUPY_AVAILABLE__ and isinstance(A_plot, cp.ndarray):
             A_plot = A_plot.get()
-        fig, ax = plt.subplots(2, 2, layout="constrained")
+        fig, ax = plt.subplots(2, 2, layout="constrained", figsize=(10, 10))
         fig.suptitle(rf"Field at $z$ = {z:.2e} m")
         ext_real = [
             self.X[0] * 1e3,
