### 2024-05-29 cfaad1c
ddgpe with new signatures
