+++ .github/workflows/tests.yml
@@ -13,6 +13,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
+          pip install -r requirements.txt
           pip install .
       - name: Test with pytest
         run: |
