+++ .github/workflows/tests.yml
@@ -10,6 +10,8 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: '3.10'
+      - name: cuda-toolkit
+        uses: Jimver/cuda-toolkit@v0.2.15
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
