+++ .github/workflows/tests.yml
@@ -2,7 +2,7 @@ on: [push]
 
 jobs:
   tests:
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-20.04
     name: Run tests
     steps:
       - uses: actions/checkout@v4
