+++ .github/workflows/tests.yml
@@ -2,7 +2,7 @@ on: [push]
 
 jobs:
   tests:
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-lastest
     name: Run tests
     steps:
       - uses: actions/checkout@v4
@@ -10,8 +10,6 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: '3.10'
-      - name: cuda-toolkit
-        uses: Jimver/cuda-toolkit@v0.2.15
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
