### 2024-05-29 e068340
try to install opencl
