+++ .github/workflows/tests.yml
@@ -12,6 +12,8 @@ jobs:
           python-version: '3.10'
       - name: Install dependencies
         run: |
+          sudo apt-get install -y ocl-icd-opencl-dev
+          sudo apt-get install -y pocl-opencl-icd
           python -m pip install --upgrade pip
           pip install -r requirements.txt
           pip install .
