### 2024-05-30 3998f7e
OpenCL installation instructions
