+++ NLSE/utils.py
@@ -13,6 +13,10 @@ except ImportError:
 
 
 try:
+    # for OpenCL backend you need to install OpenCL first
+    # sudo apt install intel-opencl-icd opencl-headers ocl-icd-opencl-dev
+    # or for AMD
+    # sudo apt install opencl-headers ocl-icd-opencl-dev
     import pyopencl
 
     __PYOPENCL_AVAILABLE__ = True
