### 2024-05-30 3200b53
Missing case in prepare_output_field
