+++ NLSE/nlse.py
@@ -260,6 +260,8 @@ class NLSE:
             integral *= c * epsilon_0 / 2
             E_00 = (self.puiss / integral) ** 0.5
             A[:] = (E_00.T * E_in.T).T
+        else:
+            A[:] = E_in
         return A, A_sq
 
     def _send_arrays_to_gpu(self) -> None:
