+++ tests/test_nlse.py
@@ -259,9 +259,18 @@ def main():
     print("Testing NLSE class")
     for backend in ["GPU", "CPU"]:
         simu = NLSE(
-            alpha, puiss, window, n2, None, L, NX=N, NY=N, Isat=Isat, backend=backend
+            alpha,
+            puiss,
+            window,
+            n2,
+            None,
+            L,
+            NX=N,
+            NY=N,
+            Isat=Isat,
+            backend=backend,
         )
-        simu.delta_z = 1e-4
+        simu.delta_z = 0.5e-4
         E_0 = np.exp(-(simu.XX**2 + simu.YY**2) / waist**2).astype(PRECISION_COMPLEX)
         simu.V = -1e-4 * np.exp(-(simu.XX**2 + simu.YY**2) / waist2**2).astype(
             PRECISION_COMPLEX
