### 2024-06-03 #6607 @taladjidi
> Hey all! I'm really sorry for the delay, but I just finished my first iteration of the review. Overall, excellent work @taladjidi! I am leaving some comments for your consideration, after which I will be happy to give this a green light -
>
>     1. Consider including a short note on the installation of `FFTW` libraries to ease the installation for `pyFFTW`. For example for my Mac machine, I had to install it via building from source.

This is now the case in the PyFFTW subsection from the README. In my case I just had to brew install it so I didn't think of it, but it was indeed not the case for some members of my team ! Thanks for the heads up.

>
>     2. Tests can be further improved -
>
>        * I might be mistaken, but `test/run_all_tests.py` doesn’t actually run the tests but rather just calls `main()` methods for all the `test_{}.py`. I couldn't see any `assert` statements in these methods.
>        * Make sure all the tests run. For example, I could not run `tests/test_broadcasting.py`.
>        * Optional - using `pytest` for the tests to have a better control over automation and assessments like coverage.

I deleted the `run_all_tests` script which was just for testing tests. There is a GitHub action that actually uses pytest to run all the tests, so this part is already covered. I improved test coverage by including array contiguity checks and memory alignment checks. However, I didn't manage to install the CUDA depedencies on the GitHub action, so for now it only checks for the CPU backend.

>     3. Consider adding `requirement.txt` for the library. Especially for building the documentation - I had to install many of the required sub-packages by `mkdocs` one-by-one.
>

Done !

>     4. Add more clear instructions to build the docs, and additionally, you may try hosting them via `Read The Docs`. Note, currently, the `Install` section is not displayed in the docs.
>

The docs now have a dedicated README file detailing build instructions, and the site should be a bit more polished now. Since it is a late addition, the site is not perfect yet but we are slowly getting there.

>     5. It would be really helpful if you could add some more `tutorials/examples` for using the library. Currently, there’s just one example in the README as a basic usage, which I feel is insufficient.
>

There is now a Jupyter notebook included in the "Tutorial" section including three different use cases and introducing the physical situation.

>        * One of these could be the code for the figures in the Paper, which could help me tick off the "reproducibility" section from my review.
>
>     6. The basic usage example doesn’t define the variable `backend`. Additionally, I could not find any docs for the visualizations that could be produced from the output one gets after the simulations. Seems like a major strong point for the library, which could be shown via examples.
>
>     7. Optional - The inheritance diagram might be adapted for dark mode on GitHub.
>

Done !

>     8. Suggestions from the paper -
>
>        * In summary - “... equations is a crucial need to model realistic experimental scenarii.” change to “... equations is crucial to modelling realistic experimental scenarios.”
>        * It might be nice to have a benchmark that reflects the mentioned “Statement of need” - to compare the computational resources for `FourierGPU.jl` and `NLSE`.
>        * Functionality - change “Graphical Processing Units” to “graphics processing units” and “Cupy(Okuta)” to “Cupy (Okuta)”.
>        * I couldn’t find any reference to `Figure 1` in the text. So you may reorder the figures accordingly. Additionally, it might be helpful for readers if you could rename the titles for sub-figures according to the caption.

I will try to address the other points today, but we are currently moving labs so this might be a bit tricky !
