### 2024-06-05 9053e7f
plotting, examples
