+++ tests/test_nonlocality.py
@@ -1,5 +1,6 @@
 from NLSE import CNLSE, CNLSE_1d, NLSE, NLSE_1d, GPE
 import numpy as np
+from scipy.constants import c, epsilon_0
 
 PRECISION_COMPLEX = np.complex64
 PRECISION_REAL = np.float32
@@ -7,7 +8,7 @@ PRECISION_REAL = np.float32
 # TODO: Add assertions to check the norm
 
 
-def main():
+def test_nonlocality():
     N = 2048
     n2 = -1.6e-9
     n12 = -1e-10
@@ -108,6 +109,12 @@ def main():
             plot=False,
             precision="single",
         )
+        arr = E.real * E.real + E.imag * E.imag
+        arr *= c * epsilon_0 / 2 * simu_c_1d.delta_X**2
+        norm = arr.sum(simu_c_1d._last_axes)
+        assert np.allclose(
+            norm, simu_c_1d.puiss
+        ), f"CNLSE_1d : Norm is not conserved ! (Backend {backend})"
         E = simu_1d.out_field(
             E_0[N // 2, :],
             L,
@@ -115,6 +122,12 @@ def main():
             plot=False,
             precision="single",
         )
+        arr = E.real * E.real + E.imag * E.imag
+        arr *= c * epsilon_0 / 2 * simu_1d.delta_X**2
+        norm = arr.sum(simu_c_1d._last_axes)
+        assert np.allclose(
+            norm, simu_1d.puiss
+        ), f"NLSE_1d : Norm is not conserved ! (Backend {backend})"
         E, V = simu_c_2d.out_field(
             np.array([E_0, V0]),
             L,
@@ -122,6 +135,12 @@ def main():
             plot=False,
             precision="single",
         )
+        arr = E.real * E.real + E.imag * E.imag
+        arr *= c * epsilon_0 / 2 * simu_c_2d.delta_X * simu_c_2d.delta_Y
+        norm = arr.sum(simu_c_2d._last_axes)
+        assert np.allclose(
+            norm, simu_c_2d.puiss
+        ), f"CNLSE : Norm is not conserved ! (Backend {backend})"
         E = simu_2d.out_field(
             E_0,
             L,
@@ -129,6 +148,12 @@ def main():
             plot=False,
             precision="single",
         )
+        arr = E.real * E.real + E.imag * E.imag
+        arr *= c * epsilon_0 / 2 * simu_2d.delta_X * simu_2d.delta_Y
+        norm = arr.sum(simu_2d._last_axes)
+        assert np.allclose(
+            norm, simu_2d.puiss
+        ), f"NLSE : Norm is not conserved ! (Backend {backend})"
         E = simu_gpe.out_field(
             E_0,
             L,
@@ -136,7 +161,9 @@ def main():
             plot=False,
             precision="single",
         )
-
-
-if __name__ == "__main__":
-    main()
+        arr = E.real * E.real + E.imag * E.imag
+        arr *= simu_gpe.delta_X * simu_gpe.delta_Y
+        norm = arr.sum(simu_gpe._last_axes)
+        assert np.allclose(
+            norm, simu_gpe.N
+        ), f"CNLSE : Norm is not conserved ! (Backend {backend})"
