### 2024-06-05 6cd6aab
No more french attribute names ...
