+++ NLSE/nlse_1d.py
@@ -16,7 +16,7 @@ class NLSE_1d(NLSE):
     def __init__(
         self,
         alpha: float,
-        puiss: float,
+        power: float,
         window: float,
         n2: float,
         V: Union[np.ndarray, None],
@@ -32,7 +32,7 @@ class NLSE_1d(NLSE):
           k0 n2 psi**2 psi
         Args:
             alpha (float): Transmission coeff
-            puiss (float): Power in W
+            power (float): Power in W
             waist (float): Waist size in m
             n2 (float): Non linear coeff in m^2/W
             V (np.ndarray) : Potential
@@ -44,7 +44,7 @@ class NLSE_1d(NLSE):
         """
         super().__init__(
             alpha=alpha,
-            puiss=puiss,
+            power=power,
             window=window,
             n2=n2,
             V=V,
@@ -83,7 +83,7 @@ class NLSE_1d(NLSE):
                 (E_in.real * E_in.real + E_in.imag * E_in.imag) * self.delta_X**2
             ).sum(axis=self._last_axes)
             integral *= c * epsilon_0 / 2
-            E_00 = (self.puiss / integral) ** 0.5
+            E_00 = (self.power / integral) ** 0.5
             A[:] = (E_00.T * E_in.T).T
         else:
             A[:] = E_in
