### 2024-06-06 e0031cc
Vortex precession
