### 2024-06-06 6162566
Benchmarks and figure code
