### 2024-06-10 #6607 @taladjidi
Hi @RMeli @obliviateandsurrender, to follow up on your comments:
- I now have created an example folder with minimum working examples of most of the software functionality
- I added two benchmarking scripts to compare with JuliaGPE.jl
- I modified the figure 1 in order to reflect these changes
- I edited the paper as you suggested
I'm currently finishing up the benchmarking runs for the figures, I feel like this would address all of the remarks you made, let me know if this is ok for you !
