### 2024-06-10 #6607 @RMeli
Thanks for the update @taladjidi. Looking at Figure 2 again, one concern I have is the variability of the GPU implementation, which spans several orders of magnitude. How do you perform the benchmark and do you know why there is such a high variability? I think it would be interesting and good to increase the total runtime (i.e. use more steps), to see if the variability decreases.
