### 2024-06-10 #6607 @taladjidi
Ah yes, sorry this was a mistake on my side ! I did not read carefully enough the docs of matplotlib :) I plotted as an error the min and max times, instead of plotting the difference between min and average ! The min and max are now within 10% of the average
