### 2024-06-10 90503f8
Paper figure updates
