+++ paper/paper.md
@@ -33,6 +33,9 @@ Providing a flexible, modern and performant framework to solve these equations i
 
 # Statement of need
 
+[`Numpy`](https://numpy.org/doc/stable/) [@harris2020array] is the default package for array operations in Python, which is the language of choice for most physicists.
+However, the performance of `Numpy` for large arrays and Fourier transforms quickly bottlenecks homemade implementations of popular solvers like the split-step spectral scheme.
+
 Over the years, there have been several packages striving to provide performant split-step solvers for NLSE type equations.
 Here are a few examples:
 
@@ -49,7 +52,7 @@ Using an easy to extend object-oriented classes, users can readily input experim
 
 $$i\partial_t\psi = -\frac{1}{2m}\nabla^2\psi + V\psi + g|\psi|^2\psi.$$
 
-In order to take advantage of the computing power of modern Graphics Processing Units (GPU) for Fast Fourier Transforms (FFT), the main workhorse of this code is the [`Cupy`](https://cupy.dev/) [@cupy_learningsys2017]  package that maps [`Numpy`](https://numpy.org/) [@harris2020array] functionalities onto the GPU using NVIDIA's [`CUDA`](https://developer.nvidia.com/cuda-downloads) API.
+In order to take advantage of the computing power of modern Graphics Processing Units (GPU) for Fast Fourier Transforms (FFT), the main workhorse of this code is the [`Cupy`](https://cupy.dev/) [@cupy_learningsys2017]  package that maps [`Numpy`](https://numpy.org/) functionalities onto the GPU using NVIDIA's [`CUDA`](https://developer.nvidia.com/cuda-downloads) API.
 It also heavily uses just-in-time compilation using [`Numba`](https://numba.pydata.org/) [@lam2015numba] in order to optimize performance while having an easily maintainable Python codebase.
 Compared to naive Numpy based CPU implementations, this package provides a 100 to 10000 times speedup for typical sizes \autoref{fig:bench}.
 While optimized for the use with GPU, it also provides a performant CPU fallback layer.
@@ -64,7 +67,12 @@ This code has been developped during the author's PhD thesis [@aladjidiFullOptic
 
 ![Example of an output of the solver. A shearing layer is observed nucleating vortices, that are attracted towards the center due to an attractive potential. The density and phase of the field are represented as well as the momentum distribution in order to get a quick overview of the state of the field.\label{fig:output}](../img/output.png)
 
-![CPU vs GPU benchmark for 1 cm of propagation (200 evolution steps).\label{fig:bench}](../img/benchmarks.png)
+![Left: CPU vs GPU vs Numpy benchmark for 1 cm of propagation (200 evolution steps).Right: Comparison versus the `JuliaGPE.jl` package on the study of vortex precession. \label{fig:bench}](../img/benchmarks.png)
+
+# Reproducibility
+
+The code used to generate the figures can be found in the [`examples`](https://github.com/Quantum-Optics-LKB/NLSE/tree/main/examples) folder of the repo with the [`fig1_turbulence.py`](https://github.com/Quantum-Optics-LKB/NLSE/blob/main/examples/fig1_turbulence.py) and [`fig2_benchmarks.py`](https://github.com/Quantum-Optics-LKB/NLSE/blob/main/examples/fig2_benchmarks.py) scripts. 
+Note that you will need a Julia install to run the `JuliaGPE.jl` script.
 
 # Acknowledgements
 
