### 2024-06-11 8453ae3
add test ddgpe + modif rabi kernel
