### 2024-06-11 9cabd59
add test ddgpe + modif rabi kernel
