+++ tests/test_ddgpe.py
@@ -1,33 +1,243 @@
 from NLSE import DDGPE
 import numpy as np
-import matplotlib.pyplot as plt
 
 if DDGPE.__CUPY_AVAILABLE__:
     import cupy as cp
-if DDGPE.__PYOPENCL_AVAILABLE__:
-    import pyopencl.array as cla
-    from pyvkfft.opencl import VkFFTApp as VkFFTApp_cl
+
 PRECISION_COMPLEX = np.complex64
 PRECISION_REAL = np.float32
 
+N = 256
+T = 1
+h_bar = 0.654  # (meV*ps)
+omega = 5.07 / h_bar  # (meV/h_bar) linear coupling (Rabi split)
+omega_exc = 1484.44 / h_bar  # (meV/h_bar) exciton energy
+omega_cav = 1482.76 / h_bar  # (meV/h_bar) cavity energy
+detuning = 0.17 / h_bar
+k_z = 27
+gamma = 0 * 0.07 / h_bar
+waist = 50
+window = 256
+g = 1e-2 / h_bar
+puiss = detuning/g
 
-def turn_on(
-    F_laser_t: np.ndarray,
-    time: np.ndarray,
-    t_up=10,
-):
-    """A function to turn on the pump more or less adiabatically
+def test_prepare_output_array() -> None:
+    for backend in ["CPU", "GPU"]:
+        simu = DDGPE(
+            gamma,
+            puiss,
+            window, 
+            g,
+            omega,
+            T, 
+            omega_exc, 
+            omega_cav, 
+            detuning, 
+            k_z, 
+            NX=N, 
+            NY=N,
+            backend=backend,
+        )
+        if backend == "CPU":
+            A = np.ones((2, N, N), dtype=PRECISION_COMPLEX)
+        elif backend == "GPU" and DDGPE.__CUPY_AVAILABLE__:
+            A = cp.ones((2, N, N), dtype=PRECISION_COMPLEX)
+        out, out_sq = simu._prepare_output_array(A, normalize=False)
+        assert (
+            out.flags.c_contiguous
+        ), f"Output array is not C-contiguous. (Backend {backend})"
+        assert (
+            out_sq.flags.c_contiguous
+        ), f"Output array is not C-contiguous. (Backend {backend})"
+        assert out.shape == (
+            2,
+            N,
+            N,
+        ), f"Output array has wrong shape. (Backend {backend})"
+        if backend == "CPU":
+            assert isinstance(
+                out, np.ndarray
+            ), f"Ouptut array type does not match backend. (Backend {backend})"
+            out /= np.max(np.abs(out))
+            A /= np.max(np.abs(A))
+            assert np.allclose(
+                out, A
+            ), f"Output array does not match input array. (Backend {backend})"
+        elif backend == "GPU" and DDGPE.__CUPY_AVAILABLE__:
+            assert isinstance(
+                out, cp.ndarray
+            ), f"Ouptut array type does not match backend. (Backend {backend})"
+            out /= cp.max(cp.abs(out))
+            A /= cp.max(cp.abs(A))
+            assert cp.allclose(
+                out, A
+            ), f"Output array does not match input array. (Backend {backend})"
 
-    Args:
-        F_laser_t (np.ndarray): self.F_pump_t as defined in class ggpe, cp.ones((int(self.t_max//self.dt)), dtype=cp.complex64)
-        time (np.ndarray):  array with the value of the time at each discretized step
-        t_up (int, optional): time taken to reach the maximum intensity (=F). Defaults to 200.
-    """
-    F_laser_t[time < t_up] = np.exp(
-        -1 * (time[time < t_up] - t_up) ** 2 / (t_up / 2) ** 2
-    )
[…]

[…]
