### 2024-06-11 57e6af3
GPU for examples, added some default callbacks
