+++ examples/nlse3d.py
@@ -35,7 +35,7 @@ def main():
         NY=N,
         NZ=NZ,
         Isat=Isat,
-        backend="CPU",
+        backend="GPU",
     )
     simu.delta_z = 0.25e-4
     E_0 = np.exp(-(simu.XX**2 + simu.YY**2) / waist**2).astype(PRECISION_COMPLEX)
