+++ examples/nonlocality.py
@@ -26,7 +26,7 @@ def main():
         NY=N,
         Isat=Isat,
         nl_length=nl_length,
-        backend="CPU",
+        backend="GPU",
     )
     simu_2d.delta_z = 1e-4
     E_0 = np.exp(-(simu_2d.XX**2 + simu_2d.YY**2) / waist**2).astype(PRECISION_COMPLEX)
