+++ examples/gpe.py
@@ -14,24 +14,22 @@ m = 87 * atomic_mass
 
 
 def main():
-    print("Testing GPE class")
-    for backend in ["CPU", "GPU"]:
-        simu_gpe = GPE(
-            gamma=0,
-            N=N_at,
-            window=window,
-            g=g,
-            V=None,
-            m=m,
-            NX=N,
-            NY=N,
-            backend=backend,
-        )
-        simu_gpe.delta_t = 1e-8
-        psi_0 = np.exp(-(simu_gpe.XX**2 + simu_gpe.YY**2) / waist**2).astype(
-            PRECISION_COMPLEX
-        )
-        simu_gpe.out_field(psi_0, 1e-6, verbose=True, plot=False, precision="single")
+    simu_gpe = GPE(
+        gamma=0,
+        N=N_at,
+        window=window,
+        g=g,
+        V=None,
+        m=m,
+        NX=N,
+        NY=N,
+        backend="GPU",
+    )
+    simu_gpe.delta_t = 1e-8
+    psi_0 = np.exp(-(simu_gpe.XX**2 + simu_gpe.YY**2) / waist**2).astype(
+        PRECISION_COMPLEX
+    )
+    simu_gpe.out_field(psi_0, 1e-6, verbose=True, plot=False, precision="single")
 
 
 if __name__ == "__main__":
