+++ examples/cnlse_1d.py
@@ -28,7 +28,7 @@ def main() -> None:
         NX=N,
         Isat=Isat,
         omega=1,
-        backend="CPU",
+        backend="GPU",
     )
     simu_c.delta_z = 1e-5
     simu_c.puiss2 = 10e-3
