+++ NLSE/callbacks.py
@@ -40,3 +40,24 @@ def norm(
     """
     if i % save_every == 0:
         norms[i // save_every] = A.real @ A.real + A.imag @ A.imag
+
+
+def evaluate_delta_n(
+    simu: NLSE, A: np.ndarray, z: float, i: int, save_every: int, delta_n: np.ndarray
+) -> None:
+    """Evaluate the non-linear refractive index change.
+
+    This will evaluate the weight of the non-linear refractive index change, allowing
+    to adjust the step size accordingly.
+
+    Args:
+        simu (NLSE): Simulation object.
+        A (np.ndarray): The current field.
+        z (float): The current propagation distance.
+        i (int): Step number.
+        save_every (int): Number of propagation steps between each step.
+        delta_n (np.ndarray): The array of delta_n values.
+    """
+    if i % save_every == 0:
+        A_sq = A.real * A.real + A.imag * A.imag
+        delta_n[i // save_every] = simu.n2 * A_sq / (1 + A_sq / simu.Isat) ** 2
