### 2024-06-12 120e770
refined logic in gpu arrays transfer
