+++ tests/test_ddgpe.py
@@ -19,22 +19,23 @@ gamma = 0 * 0.07 / h_bar
 waist = 50
 window = 256
 g = 1e-2 / h_bar
-puiss = detuning/g
+puiss = detuning / g
+
 
 def test_prepare_output_array() -> None:
     for backend in ["CPU", "GPU"]:
         simu = DDGPE(
             gamma,
             puiss,
-            window, 
+            window,
             g,
             omega,
-            T, 
-            omega_exc, 
-            omega_cav, 
-            detuning, 
-            k_z, 
-            NX=N, 
+            T,
+            omega_exc,
+            omega_cav,
+            detuning,
+            k_z,
+            NX=N,
             NY=N,
             backend=backend,
         )
@@ -73,6 +74,7 @@ def test_prepare_output_array() -> None:
                 out, A
             ), f"Output array does not match input array. (Backend {backend})"
 
+
 def test_send_arrays_to_gpu() -> None:
     if DDGPE.__CUPY_AVAILABLE__:
         omega_exc = 1484.44 / h_bar
@@ -102,10 +104,10 @@ def test_send_arrays_to_gpu() -> None:
             T,
             omega_exc,
             omega_cav,
-            detuning, 
-            k_z, 
+            detuning,
+            k_z,
             V=V,
-            NX=N, 
+            NX=N,
             NY=N,
             backend="GPU",
         )
@@ -127,9 +129,10 @@ def test_send_arrays_to_gpu() -> None:
         ), "omega cav is not a cp.ndarray. (Backend GPU)"
         assert isinstance(
             simu.omega_exc, cp.ndarray
-        ), "omega exc is not a cp.ndarray. (Backend GPU)"         
+        ), "omega exc is not a cp.ndarray. (Backend GPU)"
     else:
         pass
+
 
 def test_retrieve_arrays_from_gpu() -> None:
     if DDGPE.__CUPY_AVAILABLE__:
@@ -160,10 +163,10 @@ def test_retrieve_arrays_from_gpu() -> None:
             T,
             omega_exc,
             omega_cav,
-            detuning, 
-            k_z, 
+            detuning,
+            k_z,
             V=V,
-            NX=N, 
+            NX=N,
             NY=N,
             backend="GPU",
         )
@@ -185,25 +188,45 @@ def test_retrieve_arrays_from_gpu() -> None:
         ), "omega cav is not a np.ndarray. (Backend GPU)"
         assert isinstance(
             simu.omega_exc, np.ndarray
-        ), "omega exc is not a np.ndarray. (Backend GPU)"         
+        ), "omega exc is not a np.ndarray. (Backend GPU)"
     else:
         pass
 
+
 def test_build_propagator() -> None:
     for backend in ["GPU", "CPU"]:
-        simu = DDGPE(gamma, puiss, window, g, omega, T, omega_exc, omega_cav, detuning, k_z, NX=N, NY=N,
+        simu = DDGPE(
+            gamma,
+            puiss,
+            window,
+            g,
+            omega,
+            T,
+            omega_exc,
+            omega_cav,
+            detuning,
+            k_z,
+            NX=N,
+            NY=N,
         )
         prop = simu._build_propagator()
         assert np.allclose(
             prop[0],
-            np.exp(-1j * (simu.omega_exc * (1 + 0 * simu.Kxx**2) - simu.omega_pump) * simu.delta_z
-        )
+            np.exp(
+                -1j
+                * (simu.omega_exc * (1 + 0 * simu.Kxx**2) - simu.omega_pump)
+                * simu.delta_z
+            ),
         ), f"Propagator1 is wrong. (Backend {backend})"
         assert np.allclose(
[…]
