+++ NLSE/kernels_cl.py
@@ -27,7 +27,7 @@ def nl_prop(
     # Interactions
     arg = 1j * g * A_sq * sat
     # Losses
-    arg += -alpha / 2 * sat
+    arg += -alpha * sat
     # Potential
     arg += 1j * V
     arg *= dz
@@ -58,7 +58,7 @@ def nl_prop_without_V(
     # Interactions
     arg = 1j * g * A_sq * sat
     # Losses
-    arg += -alpha / 2 * sat
+    arg += -alpha * sat
     arg *= dz
     arg = clmath.exp(arg)
     A *= arg
@@ -93,7 +93,7 @@ def nl_prop_c(
     # Interactions
     arg = 1j * (g11 * A_sq_1 * sat + g12 * A_sq_2 * sat)
     # Losses
-    arg += -alpha / 2 * sat
+    arg += -alpha * sat
     # Potential
     arg += 1j * V
     arg *= dz
@@ -128,7 +128,7 @@ def nl_prop_without_V_c(
     # Interactions
     arg = 1j * (g11 * A_sq_1 * sat + g12 * A_sq_2 * sat)
     # Losses
-    arg += -alpha / 2 * sat
+    arg += -alpha * sat
     arg *= dz
     arg = clmath.exp(arg)
     A1 *= arg
