### 2024-06-14 Quantum-Optics-LKB/NLSE#6 @Abinashbunty
Wrong file path of inheritance & benchmarks plot in docs
This is with reference to the ongoing review of the paper in openjournals/joss-reviews#6607

The plot for benchmarks in the docs is not rendered due to wrong file path. The page is https://quantum-optics-lkb.github.io/NLSE/installation/

At the extreme end, the image should be https://github.com/Quantum-Optics-LKB/NLSE/blob/main/img/benchmarks.png
