### 2024-06-14 Quantum-Optics-LKB/NLSE#6 @Abinashbunty
In continuation, the inheritance image is also not rendered correctly due to path error. It should be https://github.com/Quantum-Optics-LKB/NLSE/blob/main/img/inheritance_graph.png on the page https://quantum-optics-lkb.github.io/NLSE/structure_and_inheritance/
