### 2024-06-14 #6607 @Abinashbunty
Hi @RMeli I just finished my review finally. Excellent package and had a great time checking several features of it and having thorough walkthrough. 😄  My earlier comment about the DOIs is fixed. Feel free to suggest otherwise. Even the tests ran on my system. 🚀  I just had few suggestions:

- The basic usage section in the docs can be extended further for users of the package. It is quite limited and might not help users to go through the usage.
- The "How to?" section is not complete. So I was wondering if that will be filled in to provide more extensive docs or will be it removed?
- 2 plots related to inheritance and benchmarks need to have the right file paths for being rendered correctly in the docs. (I have opened an issue for it in Quantum-Optics-LKB/NLSE#6

Once changes related to this are implemented, review from my side will be complete (feedback wise).
