+++ mkdocs-documentation/tests/test_cnlse.py
@@ -1,0 +1,269 @@
+from NLSE import CNLSE
+import numpy as np
+from scipy.constants import c, epsilon_0
+
+if CNLSE.__CUPY_AVAILABLE__:
+    import cupy as cp
+
+PRECISION_COMPLEX = np.complex64
+PRECISION_REAL = np.float32
+
+N = 2048
+n2 = -1.6e-9
+n12 = -1e-10
+waist = 2.23e-3
+waist2 = 70e-6
+window = 4 * waist
+power = 1.05
+Isat = 10e4  # saturation intensity in W/m^2
+Isat2 = waist / waist2 * Isat
+L = 1e-3
+alpha = 20
+
+
+def test_prepare_output_array() -> None:
+    for backend in ["CPU", "GPU"]:
+        simu = CNLSE(
+            alpha,
+            power,
+            window,
+            n2,
+            n12,
+            None,
+            L,
+            NX=N,
+            NY=N,
+            Isat=Isat,
+            backend=backend,
+        )
+        if backend == "CPU":
+            A = np.ones((2, N, N), dtype=PRECISION_COMPLEX)
+        elif backend == "GPU" and CNLSE.__CUPY_AVAILABLE__:
+            A = cp.ones((2, N, N), dtype=PRECISION_COMPLEX)
+        out, out_sq = simu._prepare_output_array(A, normalize=True)
+        assert (
+            out.flags.c_contiguous
+        ), f"Output array is not C-contiguous. (Backend {backend})"
+        assert (
+            out_sq.flags.c_contiguous
+        ), f"Output array is not C-contiguous. (Backend {backend})"
+        integral = (
+            (out.real * out.real + out.imag * out.imag) * simu.delta_X * simu.delta_Y
+        ).sum(axis=simu._last_axes)
+        integral *= c * epsilon_0 / 2
+        assert np.allclose(
+            integral,
+            np.array([simu.power, simu.puiss2]),
+            rtol=1e-4,
+        ), f"Normalization failed. (Backend {backend})"
+        assert out.shape == (
+            2,
+            N,
+            N,
+        ), f"Output array has wrong shape. (Backend {backend})"
+        if backend == "CPU":
+            assert isinstance(
+                out, np.ndarray
+            ), f"Ouptut array type does not match backend. (Backend {backend})"
+            out /= np.max(np.abs(out))
+            A /= np.max(np.abs(A))
+            assert np.allclose(
+                out, A
+            ), f"Output array does not match input array. (Backend {backend})"
+        elif backend == "GPU" and CNLSE.__CUPY_AVAILABLE__:
+            assert isinstance(
+                out, cp.ndarray
+            ), f"Ouptut array type does not match backend. (Backend {backend})"
+            out /= cp.max(cp.abs(out))
+            A /= cp.max(cp.abs(A))
+            assert cp.allclose(
+                out, A
+            ), f"Output array does not match input array. (Backend {backend})"
+
+
+def test_send_arrays_to_gpu() -> None:
+    if CNLSE.__CUPY_AVAILABLE__:
+        alpha = 20
+        Isat = 10e4
+        n2 = -1.6e-9
+        n12 = -1e-10
+        V = np.random.random((N, N)) + 1j * np.random.random((N, N))
+        alpha = np.repeat(alpha, 2)
+        alpha = alpha[..., np.newaxis, np.newaxis, np.newaxis]
+        n2 = np.repeat(n2, 2)
+        n2 = n2[..., np.newaxis, np.newaxis, np.newaxis]
+        n12 = np.repeat(n2, 2)
+        n12 = n12[..., np.newaxis, np.newaxis, np.newaxis]
+        Isat = np.repeat(Isat, 2)
[…]
