+++ mkdocs-documentation/tests/test_nlse.py
@@ -1,0 +1,254 @@
+from NLSE import NLSE
+import numpy as np
+import pyfftw
+from scipy.constants import c, epsilon_0
+
+if NLSE.__CUPY_AVAILABLE__:
+    import cupy as cp
+    from pyvkfft.cuda import VkFFTApp as VkFFTApp_cuda
+if NLSE.__PYOPENCL_AVAILABLE__:
+    import pyopencl.array as cla
+    from pyvkfft.opencl import VkFFTApp as VkFFTApp_cl
+PRECISION_COMPLEX = np.complex64
+PRECISION_REAL = np.float32
+
+
+N = 2048
+n2 = -1.6e-9
+waist = 2.23e-3
+waist2 = 70e-6
+window = 4 * waist
+power = 1.05
+Isat = 10e4  # saturation intensity in W/m^2
+L = 10e-3
+alpha = 20
+
+
+def test_build_propagator() -> None:
+    for backend in ["CPU", "GPU"]:
+        simu = NLSE(
+            alpha, power, window, n2, None, L, NX=N, NY=N, Isat=Isat, backend=backend
+        )
+        prop = simu._build_propagator()
+        assert np.allclose(
+            prop,
+            np.exp(-1j * 0.5 * (simu.Kxx**2 + simu.Kyy**2) / simu.k * simu.delta_z),
+        ), f"Propagator is wrong. (Backend {backend})"
+
+
+def test_build_fft_plan() -> None:
+    for backend in ["CPU", "GPU"]:
+        simu = NLSE(
+            alpha, power, window, n2, None, L, NX=N, NY=N, Isat=Isat, backend=backend
+        )
+        if backend == "CPU" or backend == "CL":
+            A = np.random.random((N, N)) + 1j * np.random.random((N, N))
+        elif backend == "GPU" and NLSE.__CUPY_AVAILABLE__:
+            A = cp.random.random((N, N)) + 1j * cp.random.random((N, N))
+        A = A.astype(PRECISION_COMPLEX)
+        plans = simu._build_fft_plan(A)
+        if backend == "CPU":
+            assert len(plans) == 2, f"Number of plans is wrong. (Backend {backend})"
+            assert isinstance(
+                plans[0], pyfftw.FFTW
+            ), f"Plan type is wrong. (Backend {backend})"
+            assert plans[0].output_shape == (
+                N,
+                N,
+            ), f"Plan shape is wrong. (Backend {backend})"
+        elif backend == "GPU" and NLSE.__CUPY_AVAILABLE__:
+            assert len(plans) == 1, f"Number of plans is wrong. (Backend {backend})"
+            assert isinstance(
+                plans[0], VkFFTApp_cuda
+            ), f"Plan type is wrong. (Backend {backend})"
+            assert plans[0].shape0 == (
+                N,
+                N,
+            ), f"Plan shape is wrong. (Backend {backend})"
+        elif backend == "CL" and NLSE.__PYOPENCL_AVAILABLE__:
+            assert len(plans) == 1, f"Number of plans is wrong. (Backend {backend})"
+            assert isinstance(
+                plans[0], VkFFTApp_cl
+            ), f"Plan type is wrong. (Backend {backend})"
+            assert plans[0].shape0 == (
+                N,
+                N,
+            ), f"Plan shape is wrong. (Backend {backend})"
+
+
+def test_prepare_output_array() -> None:
+    for backend in ["CPU", "GPU"]:
+        simu = NLSE(
+            alpha, power, window, n2, None, L, NX=N, NY=N, Isat=Isat, backend=backend
+        )
+        if backend == "CPU" or backend == "CL":
+            A = np.random.random((N, N)) + 1j * np.random.random((N, N))
+        elif backend == "GPU" and NLSE.__CUPY_AVAILABLE__:
[…]
