+++ mkdocs-documentation/docs/nlse.md
@@ -1,0 +1,5 @@
+In this page you will find the reference documentation for the NLSE class.
+
+::: NLSE.nlse.NLSE
+    options:
+      show_source: true
