+++ mkdocs-documentation/docs/cnlse_1d.md
@@ -1,0 +1,5 @@
+In this page you will find the reference documentation for the CNLSE_1d class.
+
+::: NLSE.cnlse_1d.CNLSE_1d
+    options:
+      show_source: true
