+++ mkdocs-documentation/docs/ddgpe.md
@@ -1,0 +1,5 @@
+In this page you will find the reference documentation for the GPE class.
+
+::: NLSE.ddgpe.DDGPE
+    options:
+      show_source: true
