+++ mkdocs-documentation/docs/gpe.md
@@ -1,0 +1,5 @@
+In this page you will find the reference documentation for the CNLSE class.
+
+::: NLSE.gpe.GPE
+    options:
+      show_source: true
