+++ mkdocs-documentation/docs/installation.md
@@ -24,7 +24,7 @@ This code has been tested on the three main platforms: Linux, MacOs and Windows.
 
 For optimal speed, this code uses your GPU (graphics card). For this, you need specific libraries. For Nvidia cards, you need a [CUDA](https://developer.nvidia.com/cuda-toolkit) install. For AMD cards, you need a [ROCm](https://rocmdocs.amd.com/en/latest/) install. Of course, you need to update your graphics driver to take full advantage of these. In any case we use [CuPy](https://cupy.dev) for the Python interface to these libraries.
 
-**The `cupy` dependency is not included in [`setup.py`](/setup.py) in order to not break installation on platforms that do not support it !**
+**The `cupy` dependency is not included in [`setup.py`](https://github.com/Quantum-Optics-LKB/NLSE/tree/main/setup.py) in order to not break installation on platforms that do not support it !**
 
 ### PyFFTW
 
@@ -46,7 +46,7 @@ Tests are included to check functionalities and benchmark performance.
 You can run all tests by executing `pytest` at the root of the package (warning: this might take some time !).
 It will test both CPU and GPU backends.
 
-The benchmarks can be run using [`tests/benchmarks.py`](/tests/benchmarks.py) and compare a "naive" numpy implementation of the main solver loop to our solver.
+The benchmarks can be run using [`tests/benchmarks.py`](https://github.com/Quantum-Optics-LKB/NLSE/tree/main/tests/benchmarks.py) and compare a "naive" numpy implementation of the main solver loop to our solver.
 On a Nvidia RTX4090 GPU and Ryzen 7950X CPU, we test our solver to the following results:
-![benchmarks](/img/benchmarks.png)
+![benchmarks](https://github.com/Quantum-Optics-LKB/NLSE/tree/main/img/benchmarks.png)
 
