+++ NLSE/kernels_gpu.py
@@ -168,7 +168,7 @@ def vortex_cp(
         j (int): position column of the vortex
         ii (int): meshgrid position row (coordinates of the image)
         jj (int): meshgrid position column (coordinates of the image)
-        l (int): vortex charge
+        ll (int): vortex charge
 
     Returns:
         None
