### 2024-06-24 #6607 @taladjidi
Hi @Abinashbunty thanks for your comments.
- I kept the "basic usage" section real simple since it's meant as a minimum working example, however I did make it a tiny bit nicer
- I overhauled the "how to" section
- The links are fixed, working with mkdocs was a bit finicky but I managed to make it work. I closed your issue concerning this point.

I also added the full reference to all the classes, and fixed the docstrings so that now everything gets parsed correctly.
