### 2024-07-03 3c3145c
Bad condition in test_ddgpe and asymetric window
