+++ paper/paper.bib
@@ -68,6 +68,7 @@
   PDF = {https://theses.hal.science/tel-04391272/file/ALADJIDI_Tangui_these_2023.pdf},
   HAL_ID = {tel-04391272},
   HAL_VERSION = {v1},
+  doi = {10.5281/zenodo.12698001}
 }
 
 @article{glorieuxHotAtomicVapors2023,
@@ -131,6 +132,7 @@
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
-  year={2015}
+  year={2015},
+  doi = {10.1145/2833157.2833162}
 }
 
