### 2024-07-16 #6607 @RMeli
@taladjidi I had a look into the missing DOIs and I think they are not available. Can you please confirm that? Thank you.
