### 2024-07-16 #6607 @taladjidi
Hi @RMeli , the DOI for the Pitaevskii Stringari works fine on my end, and there is no DOI for Cupy. I did use the official citation provided on their repo though.
