### 2024-07-16 #6607 @RMeli
That's right, the suggested DOI `10.1017/cbo9780511524240.005` does seem to work. I missed this. Can you please add it to the `paper.bib`?

https://github.com/Quantum-Optics-LKB/NLSE/blob/138f6a8b63f4f0e5a7c0e5c061fb3da328ca531d/paper/paper.bib#L1-L11
