+++ paper/paper.bib
@@ -7,7 +7,8 @@
   number    = {164},
   publisher = {Oxford University Press},
   author    = {Pitaevskij, Lev P. and Stringari, Sandro},
-  year      = {2016}
+  year      = {2016},
+  doi = {10.1017/cbo9780511524240.005}
 }
 
 @article{Schloss2018, 
