+++ NLSE/cnlse_1d.py
@@ -82,11 +82,11 @@ class CNLSE_1d(CNLSE):
             A = cp.empty_like(E)
             A_sq = cp.empty_like(E, dtype=E.real.dtype)
             E = cp.asarray(E)
-            puiss_arr = cp.array([self.power, self.puiss2], dtype=E.dtype)
+            puiss_arr = cp.array([self.power, self.power2], dtype=E.dtype)
         else:
             A = pyfftw.empty_aligned(E.shape, dtype=E.dtype)
             A_sq = np.empty_like(E, dtype=E.real.dtype)
-            puiss_arr = np.array([self.power, self.puiss2], dtype=E.dtype)
+            puiss_arr = np.array([self.power, self.power2], dtype=E.dtype)
         if normalize:
             # normalization of the field
             integral = ((E.real * E.real + E.imag * E.imag) * self.delta_X**2).sum(
