### 2024-07-22 #6607 @taladjidi
Hi @RMeli , I uploaded the archive of the last release on [Zenodo](https://doi.org/10.5281/zenodo.12795396).
I also changed the license to stronger GNUGPLv3 license. I hope this is OK.
