### 2024-07-22 #6607 @RMeli
> I also changed the license to stronger GNUGPLv3 license. I hope this is OK.

Yes, this is an OSI-approved license, therefore it is OK.
