### 2024-07-22 #6607 @RMeli
@taladjidi can you please change the title of the Zenodo archive to match the title of the paper (i.e. remove the version, which is already accounted for by the DOI)? Also, your two co-authors appear as "Contact Persons" and not as authors on Zenodo. Can you please make sure they are in the correct field?
