### 2024-07-22 #6607 @taladjidi
> @taladjidi can you please change the title of the Zenodo archive to match the title of the paper (i.e. remove the version, which is already accounted for by the DOI)? Also, your two co-authors appear as "Contact Persons" and not as authors on Zenodo. Can you please make sure they are in the correct field?

Done !

> @taladjidi can you please update the Zenodo metadata to include the full author affiliation (as in the paper)? Thank you.

The lab appears as a ROR (research organization registry) entry already, isn't it correct ? When you click on the ROR logo you get the same affiliation as the paper.
