### 2024-07-22 #6607 @RMeli
> The lab appears as a ROR (research organization registry) entry already, isn't it correct ?

Ah! Thanks for pointing this out. I was just hovering over the author name to see the affiliation. Using "Show affiliation" does indeed show the ROR logo. However, the last author has twice the same org (one of which is not a ROR entry). Would you be able to remove it? Then I think we are good to go. ;)
