### 2024-07-23 #6607 @crvernon
:wave: @taladjidi - great job on this!

I just need you to address the following before I move to accept this one for publication:

- [ ] Please ensure the title of your Zenodo archive ("NLSE: A python package to solve the non-linear Schrödinger equation") matches the title of your paper ("NLSE: A Python package to solve the nonlinear Schrödinger equation") exactly.  Feel free to edit which ever you like to make them match.  No need to conduct another release.

- [ ] LINE 7-8:  The phrase "...equation allowing to model the propagation of light..." is missing something.  Perhaps it should be stated as "...equation used to model the propagation of light..."
- [ ] LINE 16:  "Numpy" should have a capital "P" in this context to appear as "NumPy"
- [ ] LINE 17:  "Numpy" should have a capital "P" in this context to appear as "NumPy"
- [ ] LINE 28-29: "...Using an easy to extend object-oriented classes, users..." should read as the following:  "...Using easy to extend object-oriented classes, users..." with the removal of "an"
- [ ] LINE 34:  "Cupy" in this context should have a capital "P" and be written as "CuPY"
- [ ] LINE 35:  "Numpy" should have a capital "P" in this context to appear as "NumPy"
- [ ] LINE 37:  "Numpy" should have a capital "P" in this context to appear as "NumPy"
- [ ] Figure 2 appears in-text before Figure 1.  Please configure these so that they appear in order by making the performance figure appear first as Figure 1.
- [ ] LINE 44:  "Numpy" should have a capital "P" in this context to appear as "NumPy"
- [ ] LINE 47:  "This code has been developed initially developed in..." should be written as "This code was initially developed in..."
- [ ] In the Figure 2 caption, there is a missing space after the period in "...(200 evolution steps).Right..."
- [ ] LINE 52-53:  "...you will need a Julia install to run..." should read "...you will need Julia installed to run..."
- [ ] LINE 75:  "llvm" and "jit" should be written in all capital letters.  You can maintain capitalization in your references by placing any content you wish to maintain formatting for in curly brackets in your bib file.
- [ ] LINE 85:  The "e" in "einstein" should be capitalized
- [ ] LINE 87:  The "g" and "p" in "gross--pitaevskii" should be capitalized

Let me know when these changes have been addressed.
