+++ paper/paper.md
@@ -26,15 +26,15 @@ bibliography: paper.bib
 
 # Summary
 
-The non-linear Schrödinger equation (NLSE) is a general non-linear equation allowing to model the propagation of light in non-linear media.
+The non-linear Schrödinger equation (NLSE) is a general non-linear equation used to model the propagation of light in non-linear media.
 This equation is mathematically isomorphic to the Gross-Pitaevskii equation (GPE) [@pitaevskij_bose-einstein_2016] describing the evolution of cold atomic ensembles.
 Recently, the growing field of quantum fluids of light [@carusotto_quantum_2013] has proven a fruitful testbed for several fundamental quantum and classical phenomena such as superfluidity [@michel_superfluid_2018] or turbulence [@bakerrasooliTurbulentDynamicsTwodimensional2023].
 Providing a flexible, modern and performant framework to solve these equations is crucial to model realistic experimental scenarios.
 
 # Statement of need
 
-[`Numpy`](https://numpy.org/doc/stable/) [@harris2020array] is the default package for array operations in Python, which is the language of choice for most physicists.
-However, the performance of `Numpy` for large arrays and Fourier transforms quickly bottlenecks homemade implementations of popular solvers like the split-step spectral scheme.
+[`NumPy`](https://numpy.org/doc/stable/) [@harris2020array] is the default package for array operations in Python, which is the language of choice for most physicists.
+However, the performance of `NumPy` for large arrays and Fourier transforms quickly bottlenecks homemade implementations of popular solvers like the split-step spectral scheme.
 
 Over the years, there have been several packages striving to provide performant split-step solvers for NLSE type equations.
 Here are a few examples:
@@ -44,7 +44,7 @@ Here are a few examples:
 - [`py-fmas`](https://github.com/omelchert/py-fmas) for 1D NLSE in optical fibers, with a split-step method (currently unmaintained).
 
 With our project, we bring similar performance to C++ and Julia implementations, while striving for accessibility and maintainability by using Python.
-Using an easy to extend object-oriented classes, users can readily input experimental parameters to quickly model real setups.
+Using easy to extend object-oriented classes, users can readily input experimental parameters to quickly model real setups.
 
 # Functionality
 
@@ -52,27 +52,27 @@ Using an easy to extend object-oriented classes, users can readily input experim
 
 $$i\partial_t\psi = -\frac{1}{2m}\nabla^2\psi + V\psi + g|\psi|^2\psi.$$
 
-To take advantage of the computing power of modern Graphics Processing Units (GPUs) for Fast Fourier Transforms (FFTs), the main workhorse of this code is the [`Cupy`](https://cupy.dev/) [@cupy_learningsys2017]  package that maps [`Numpy`](https://numpy.org/) functionalities onto the GPU using NVIDIA's [`CUDA`](https://developer.nvidia.com/cuda-downloads) API.
+To take advantage of the computing power of modern Graphics Processing Units (GPUs) for Fast Fourier Transforms (FFTs), the main workhorse of this code is the [`CuPy`](https://cupy.dev/) [@cupy_learningsys2017]  package that maps [`NumPy`](https://numpy.org/) functionalities onto the GPU using NVIDIA's [`CUDA`](https://developer.nvidia.com/cuda-downloads) API.
 It also heavily uses just-in-time compilation using [`Numba`](https://numba.pydata.org/) [@lam2015numba] to optimize performance while having an easily maintainable Python codebase.
-Compared to naive Numpy-based CPU implementations, this package provides a 100 to 10000 times speedup for typical sizes \autoref{fig:bench}.
[…]
