+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @book{pitaevskij_bose-einstein_2016,
   location  = {Oxford},
-  title     = {Bose-Einstein condensation and superfluidity},
+  title     = {{Bose-Einstein} condensation and superfluidity},
   isbn      = {978-0-19-875888-4},
   series    = {International series of monographs on physics},
   pagetotal = {553},
@@ -20,7 +20,7 @@
   number = {32}, 
   pages = {1037}, 
   author = {James R. Schloss and Lee J. O'Riordan}, 
-  title = {GPUE: Graphics Processing Unit Gross--Pitaevskii Equation solver}, 
+  title = {{GPUE: Graphics Processing Unit Gross--Pitaevskii Equation solver}}, 
   journal = {Journal of Open Source Software} 
 } 
 
@@ -129,7 +129,7 @@
 }
 
 @inproceedings{lam2015numba,
-  title={Numba: A llvm-based python jit compiler},
+  title={Numba: A {LLVM}-based python {JIT} compiler},
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
