+++ paper/paper.bib
@@ -20,7 +20,7 @@
   number = {32}, 
   pages = {1037}, 
   author = {James R. Schloss and Lee J. O'Riordan}, 
-  title = {{GPUE: Graphics Processing Unit Gross--Pitaevskii Equation solver}}, 
+  title = {{GPUE: Graphics Processing Unit Gross-Pitaevskii Equation solver}}, 
   journal = {Journal of Open Source Software} 
 } 
 
