### 2024-07-23 #6607 @taladjidi
Hi @crvernon, thanks for the swift takeover ! I just addressed everything you mentioned. I edited all the "nonlinear" without the hyphen to be more consistent with the title. Thanks for the .bib tips I didn't know !
