### 2024-07-23 #6607 @crvernon
Thanks @taladjidi!

I still need you to:

- [ ] Capitalize the "p" in "python" in your Zenodo archive title

and address:
- [ ]  Figure 2 appears in-text before Figure 1. Please configure these so that they appear in order by making the performance figure appear first as Figure 1.
