### 2024-07-23 c25d85e
Update README.md with JOSS logo
